\name{acc_LOBOV}
\alias{acc_LOBOV}
\docType{data}
\title{
Accuracies of the LOBOV
}
\description{
Accuracies of the Leave One Biobank Validation calculated in Bizzarri et al.
}
\usage{data("acc_LOBOV")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\examples{
data(acc_LOBOV)
## maybe str(acc_LOBOV) ; plot(acc_LOBOV) ...
}
\keyword{datasets}
