% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{activateButtn}
\alias{activateButtn}
\title{activateButtn}
\usage{
activateButtn(check1, check2, button)
}
\value{
button activation
}
\description{
helper function to activate buttons based on 2 checks
}
\keyword{internal}
