% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Ahola_Olli_betas}
\alias{Ahola_Olli_betas}
\title{T2D-score Betas}
\format{
An object of class \code{data.frame} with 7 rows and 3 columns.
}
\usage{
data("Ahola_Olli_betas")
}
\description{
The coefficients used to compute the T2Diabetes score by Ahola Olli.
}
\details{
Dataframe containing the abbreviation of the metabolites, the metabolites names and finally the Coefficients to compute the T2Diabetes score
}
\examples{
data("Ahola_Olli_betas")

}
\references{
Ahola-Olli,A.V. et al. (2019) Circulating metabolites and the risk of type 2 diabetes: a prospective study of 11,896 young adults from four Finnish cohorts. Diabetologia, 62, 2298-2309, \url{doi:10.1007/s00125-019-05001-w}
}
\keyword{datasets}
