% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{apply.fit}
\alias{apply.fit}
\title{Function that apply the MetaboAge model to the NH-metabolomics concentrations}
\usage{
apply.fit(mat, FIT)
}
\arguments{
\item{mat}{The NH-metabolomics matrix}

\item{FIT}{The betas of the linear regression composing the MetaboAge by van den Akker et al.}
}
\value{
The MetaboAge by van den Akker et al.
}
\description{
Function that apply the MetaboAge model to the NH-metabolomics concentrations
}
