\name{BBMRI_hist}
\alias{BBMRI_hist}
\docType{data}
\title{
Histograms of the metabolites in BBMRI-nl
}
\description{
This list collects the histograms of the 57 metabolites selected collected in BBMRI-nl with which the user will be able to compare his/her own dataset.
}
\usage{data("BBMRI_hist")}
\format{
  The format is:
List of 57
 $ ala       :List of 6
  ..$ breaks  : num [1:100] 0.0691 0.0764 0.0836 0.0908 0.0981 ...
  ..$ counts  : int [1:99] 1 0 1 2 1 4 11 18 46 49 ...
  ..$ density : num [1:99] 0.00443 0 0.00443 0.00885 0.00443 ...
  ..$ mids    : num [1:99] 0.0727 0.08 0.0872 0.0945 0.1017 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ gln       :List of 6
  ..$ breaks  : num [1:100] 0.0406 0.0495 0.0585 0.0674 0.0764 ...
  ..$ counts  : int [1:99] 2 0 1 0 0 1 2 0 1 3 ...
  ..$ density : num [1:99] 0.00834 0 0.00417 0 0 ...
  ..$ mids    : num [1:99] 0.045 0.054 0.0629 0.0719 0.0808 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ his       :List of 6
  ..$ breaks  : num [1:100] 1.96e-06 6.39e-03 1.28e-02 1.92e-02 2.55e-02 ...
  ..$ counts  : int [1:99] 7 16 84 299 855 1907 3387 4752 5676 5354 ...
  ..$ density : num [1:99] 0.0352 0.0804 0.4219 1.5018 4.2945 ...
  ..$ mids    : num [1:99] 0.0032 0.00958 0.01597 0.02236 0.02874 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ phe       :List of 6
  ..$ breaks  : num [1:100] 0.016 0.0184 0.0208 0.0231 0.0255 ...
  ..$ counts  : int [1:99] 1 3 5 45 153 455 940 1814 2714 3292 ...
  ..$ density : num [1:99] 0.0136 0.0408 0.0681 0.6126 2.0827 ...
  ..$ mids    : num [1:99] 0.0172 0.0196 0.0219 0.0243 0.0267 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ tyr       :List of 6
  ..$ breaks  : num [1:100] 0.0167 0.0181 0.0196 0.0211 0.0226 ...
  ..$ counts  : int [1:99] 1 3 7 8 15 28 40 46 80 113 ...
  ..$ density : num [1:99] 0.0218 0.0655 0.1527 0.1745 0.3273 ...
  ..$ mids    : num [1:99] 0.0174 0.0189 0.0204 0.0218 0.0233 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ ile       :List of 6
  ..$ breaks  : num [1:100] 0.000928 0.00299 0.005051 0.007112 0.009174 ...
  ..$ counts  : int [1:99] 1 0 0 3 9 19 35 58 105 168 ...
  ..$ density : num [1:99] 0.0156 0 0 0.0467 0.14 ...
  ..$ mids    : num [1:99] 0.00196 0.00402 0.00608 0.00814 0.0102 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ leu       :List of 6
  ..$ breaks  : num [1:100] 0 0.00215 0.0043 0.00645 0.00861 ...
  ..$ counts  : int [1:99] 1 0 0 0 1 0 1 1 4 6 ...
  ..$ density : num [1:99] 0.0149 0 0 0 0.0149 ...
  ..$ mids    : num [1:99] 0.00108 0.00323 0.00538 0.00753 0.00968 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ val       :List of 6
  ..$ breaks  : num [1:100] 0.00795 0.01257 0.01719 0.0218 0.02642 ...
  ..$ counts  : int [1:99] 2 2 3 2 3 1 5 5 6 9 ...
  ..$ density : num [1:99] 0.0139 0.0139 0.0209 0.0139 0.0209 ...
  ..$ mids    : num [1:99] 0.0103 0.0149 0.0195 0.0241 0.0287 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ glc       :List of 6
  ..$ breaks  : num [1:100] 0 0.223 0.445 0.668 0.89 ...
  ..$ counts  : int [1:99] 88 50 56 56 69 69 92 98 108 143 ...
  ..$ density : num [1:99] 0.01271 0.00722 0.00809 0.00809 0.00996 ...
  ..$ mids    : num [1:99] 0.111 0.334 0.556 0.779 1.001 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ lac       :List of 6
  ..$ breaks  : num [1:100] 0.153 0.301 0.448 0.595 0.743 ...
  ..$ counts  : int [1:99] 85 361 1837 4156 5210 4875 3994 2944 1972 1342 ...
  ..$ density : num [1:99] 0.0185 0.0785 0.3995 0.9038 1.133 ...
  ..$ mids    : num [1:99] 0.227 0.374 0.522 0.669 0.816 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ cit       :List of 6
  ..$ breaks  : num [1:100] 1.32e-05 1.90e-01 3.81e-01 5.71e-01 7.61e-01 ...
  ..$ counts  : int [1:99] 30968 222 2 0 0 0 1 1 0 0 ...
  ..$ density : num [1:99] 5.216041 0.037392 0.000337 0 0 ...
  ..$ mids    : num [1:99] 0.0952 0.2855 0.4758 0.6661 0.8564 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ ace       :List of 6
  ..$ breaks  : num [1:100] 0.00397 0.02377 0.04357 0.06336 0.08316 ...
  ..$ counts  : int [1:99] 1542 13598 9185 4641 1479 360 107 38 20 12 ...
  ..$ density : num [1:99] 2.5 22.06 14.9 7.53 2.4 ...
  ..$ mids    : num [1:99] 0.0139 0.0337 0.0535 0.0733 0.0931 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ acace     :List of 6
  ..$ breaks  : num [1:100] 0 0.0111 0.0223 0.0334 0.0446 ...
  ..$ counts  : int [1:99] 1213 5277 8407 6427 3659 2176 1201 710 502 337 ...
  ..$ density : num [1:99] 3.51 15.27 24.33 18.6 10.59 ...
  ..$ mids    : num [1:99] 0.00557 0.01671 0.02785 0.03899 0.05014 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ crea      :List of 6
  ..$ breaks  : num [1:100] 0.0161 0.0252 0.0342 0.0432 0.0523 ...
  ..$ counts  : int [1:99] 3 24 460 3132 7060 7698 5700 3338 1650 831 ...
  ..$ density : num [1:99] 0.0107 0.0856 1.641 11.1731 25.1858 ...
  ..$ mids    : num [1:99] 0.0207 0.0297 0.0387 0.0478 0.0568 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ alb       :List of 6
  ..$ breaks  : num [1:100] 0.0108 0.0121 0.0134 0.0147 0.0161 ...
  ..$ counts  : int [1:99] 1 0 0 0 0 0 0 0 0 0 ...
  ..$ density : num [1:99] 0.0245 0 0 0 0 ...
  ..$ mids    : num [1:99] 0.0115 0.0128 0.0141 0.0154 0.0167 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ gp        :List of 6
  ..$ breaks  : num [1:100] 0.627 0.659 0.691 0.724 0.756 ...
  ..$ counts  : int [1:99] 2 2 1 2 3 11 27 37 58 159 ...
  ..$ density : num [1:99] 0.001985 0.001985 0.000993 0.001985 0.002978 ...
  ..$ mids    : num [1:99] 0.643 0.675 0.707 0.74 0.772 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ m_vldl_l  :List of 6
  ..$ breaks  : num [1:100] 0 0.0623 0.1246 0.1869 0.2492 ...
  ..$ counts  : int [1:99] 34 127 655 1602 2494 3010 3000 2894 2627 2267 ...
  ..$ density : num [1:99] 0.0175 0.0653 0.3368 0.8237 1.2824 ...
  ..$ mids    : num [1:99] 0.0312 0.0935 0.1558 0.2181 0.2804 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ s_vldl_l  :List of 6
  ..$ breaks  : num [1:100] 0 0.0369 0.0738 0.1108 0.1477 ...
  ..$ counts  : int [1:99] 19 0 1 6 28 92 217 346 585 790 ...
  ..$ density : num [1:99] 0.016486 0 0.000868 0.005206 0.024295 ...
  ..$ mids    : num [1:99] 0.0185 0.0554 0.0923 0.1292 0.1661 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ xs_vldl_l :List of 6
  ..$ breaks  : num [1:100] 0 0.0176 0.0351 0.0527 0.0702 ...
  ..$ counts  : int [1:99] 10 0 0 0 0 0 0 6 9 19 ...
  ..$ density : num [1:99] 0.0182 0 0 0 0 ...
  ..$ mids    : num [1:99] 0.00878 0.02633 0.04389 0.06144 0.079 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ idl_l     :List of 6
  ..$ breaks  : num [1:100] 0.0518 0.0914 0.1311 0.1707 0.2104 ...
  ..$ counts  : int [1:99] 1 0 4 4 16 34 33 62 58 99 ...
  ..$ density : num [1:99] 0.000808 0 0.003231 0.003231 0.012924 ...
  ..$ mids    : num [1:99] 0.0716 0.1112 0.1509 0.1906 0.2302 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ l_ldl_l   :List of 6
  ..$ breaks  : num [1:100] 0 0.0502 0.1004 0.1506 0.2008 ...
  ..$ counts  : int [1:99] 32 0 1 2 18 30 58 119 140 261 ...
  ..$ density : num [1:99] 0.020418 0 0.000638 0.001276 0.011485 ...
  ..$ mids    : num [1:99] 0.0251 0.0753 0.1255 0.1757 0.226 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ m_ldl_l   :List of 6
  ..$ breaks  : num [1:100] 0 0.0294 0.0588 0.0882 0.1175 ...
  ..$ counts  : int [1:99] 134 0 0 0 3 46 101 151 260 373 ...
  ..$ density : num [1:99] 0.14609 0 0 0 0.00327 ...
  ..$ mids    : num [1:99] 0.0147 0.0441 0.0735 0.1028 0.1322 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ s_ldl_l   :List of 6
  ..$ breaks  : num [1:100] 0 0.0179 0.0358 0.0537 0.0716 ...
  ..$ counts  : int [1:99] 134 0 0 0 3 8 42 69 135 201 ...
  ..$ density : num [1:99] 0.23982 0 0 0 0.00537 ...
  ..$ mids    : num [1:99] 0.00895 0.02685 0.04475 0.06265 0.08055 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ m_hdl_l   :List of 6
  ..$ breaks  : num [1:100] 0 0.0184 0.0368 0.0552 0.0737 ...
  ..$ counts  : int [1:99] 345 0 0 0 0 0 1 0 0 0 ...
  ..$ density : num [1:99] 0.6 0 0 0 0 ...
  ..$ mids    : num [1:99] 0.00921 0.02762 0.04604 0.06445 0.08286 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ s_hdl_l   :List of 6
  ..$ breaks  : num [1:100] 0 0.0179 0.0358 0.0536 0.0715 ...
  ..$ counts  : int [1:99] 39 0 0 0 0 0 0 0 0 0 ...
  ..$ density : num [1:99] 0.0699 0 0 0 0 ...
  ..$ mids    : num [1:99] 0.00894 0.02682 0.0447 0.06258 0.08045 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ idl_c     :List of 6
  ..$ breaks  : num [1:100] 0 0.0272 0.0545 0.0817 0.1089 ...
  ..$ counts  : int [1:99] 33 3 8 7 25 33 51 58 81 127 ...
  ..$ density : num [1:99] 0.03883 0.00353 0.00941 0.00824 0.02942 ...
  ..$ mids    : num [1:99] 0.0136 0.0408 0.0681 0.0953 0.1225 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ serum_c   :List of 6
  ..$ breaks  : num [1:100] 0.371 0.515 0.66 0.804 0.948 ...
  ..$ counts  : int [1:99] 1 0 2 8 19 20 39 38 50 66 ...
  ..$ density : num [1:99] 0.000222 0 0.000445 0.001778 0.004223 ...
  ..$ mids    : num [1:99] 0.443 0.587 0.732 0.876 1.02 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ vldl_c    :List of 6
  ..$ breaks  : num [1:100] 1.96e-09 4.60e-02 9.19e-02 1.38e-01 1.84e-01 ...
  ..$ counts  : int [1:99] 1 4 18 47 111 169 291 399 558 772 ...
  ..$ density : num [1:99] 0.000697 0.002787 0.012544 0.032753 0.077352 ...
  ..$ mids    : num [1:99] 0.023 0.069 0.115 0.161 0.207 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ ldl_c     :List of 6
  ..$ breaks  : num [1:100] 6.34e-13 7.03e-02 1.41e-01 2.11e-01 2.81e-01 ...
  ..$ counts  : int [1:99] 8 23 29 56 104 141 237 301 413 598 ...
  ..$ density : num [1:99] 0.00365 0.01048 0.01322 0.02552 0.0474 ...
  ..$ mids    : num [1:99] 0.0351 0.1054 0.1757 0.246 0.3163 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ hdl_c     :List of 6
  ..$ breaks  : num [1:100] 0.0589 0.0903 0.1218 0.1532 0.1847 ...
  ..$ counts  : int [1:99] 1 0 1 1 1 6 3 5 10 16 ...
  ..$ density : num [1:99] 0.00102 0 0.00102 0.00102 0.00102 ...
  ..$ mids    : num [1:99] 0.0746 0.1061 0.1375 0.169 0.2004 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ hdl2_c    :List of 6
  ..$ breaks  : num [1:100] 0 0.0264 0.0528 0.0792 0.1057 ...
  ..$ counts  : int [1:99] 124 32 47 53 46 80 82 83 101 128 ...
  ..$ density : num [1:99] 0.1504 0.0388 0.057 0.0643 0.0558 ...
  ..$ mids    : num [1:99] 0.0132 0.0396 0.066 0.0924 0.1189 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ hdl3_c    :List of 6
  ..$ breaks  : num [1:100] 0.0589 0.0687 0.0786 0.0884 0.0983 ...
  ..$ counts  : int [1:99] 1 0 0 0 0 0 1 0 0 0 ...
  ..$ density : num [1:99] 0.00325 0 0 0 0 ...
  ..$ mids    : num [1:99] 0.0638 0.0737 0.0835 0.0934 0.1032 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ vldl_d    :List of 6
  ..$ breaks  : num [1:100] 32.6 32.7 32.8 33 33.1 ...
  ..$ counts  : int [1:99] 1 0 1 1 3 1 5 13 13 14 ...
  ..$ density : num [1:99] 0.000277 0 0.000277 0.000277 0.000831 ...
  ..$ mids    : num [1:99] 32.7 32.8 32.9 33 33.1 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ ldl_d     :List of 6
  ..$ breaks  : num [1:100] 21.9 21.9 22 22 22.1 ...
  ..$ counts  : int [1:99] 2 1 1 0 0 0 0 2 0 0 ...
  ..$ density : num [1:99] 0.001767 0.000883 0.000883 0 0 ...
  ..$ mids    : num [1:99] 21.9 22 22 22 22.1 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ hdl_d     :List of 6
  ..$ breaks  : num [1:100] 8.98 9 9.02 9.04 9.06 ...
  ..$ counts  : int [1:99] 1 0 0 0 0 1 3 2 1 3 ...
  ..$ density : num [1:99] 0.0016 0 0 0 0 ...
  ..$ mids    : num [1:99] 8.99 9.01 9.03 9.05 9.07 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ serum_tg  :List of 6
  ..$ breaks  : num [1:100] 0.117 0.235 0.353 0.472 0.59 ...
  ..$ counts  : int [1:99] 6 61 308 944 1926 2894 3040 3214 2892 2666 ...
  ..$ density : num [1:99] 0.00163 0.01653 0.08347 0.25582 0.52194 ...
  ..$ mids    : num [1:99] 0.176 0.294 0.413 0.531 0.649 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ totpg     :List of 6
  ..$ breaks  : num [1:100] 0.194 0.238 0.281 0.325 0.368 ...
  ..$ counts  : int [1:99] 2 0 2 3 5 2 8 30 25 35 ...
  ..$ density : num [1:99] 0.00148 0 0.00148 0.00222 0.00369 ...
  ..$ mids    : num [1:99] 0.216 0.259 0.303 0.346 0.39 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ pc        :List of 6
  ..$ breaks  : num [1:100] 0 0.0498 0.0995 0.1493 0.199 ...
  ..$ counts  : int [1:99] 14 0 0 0 1 0 4 6 4 25 ...
  ..$ density : num [1:99] 0.009154 0 0 0 0.000654 ...
  ..$ mids    : num [1:99] 0.0249 0.0746 0.1244 0.1742 0.2239 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ sm        :List of 6
  ..$ breaks  : num [1:100] 0.128 0.138 0.149 0.159 0.17 ...
  ..$ counts  : int [1:99] 1 1 0 5 10 8 13 25 30 42 ...
  ..$ density : num [1:99] 0.00305 0.00305 0 0.01525 0.03051 ...
  ..$ mids    : num [1:99] 0.133 0.143 0.154 0.164 0.175 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ totcho    :List of 6
  ..$ breaks  : num [1:100] 0.0621 0.1132 0.1644 0.2155 0.2667 ...
  ..$ counts  : int [1:99] 1 0 0 0 0 0 0 0 2 2 ...
  ..$ density : num [1:99] 0.000628 0 0 0 0 ...
  ..$ mids    : num [1:99] 0.0877 0.1388 0.19 0.2411 0.2922 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ apoa1     :List of 6
  ..$ breaks  : num [1:100] 0.545 0.566 0.586 0.607 0.628 ...
  ..$ counts  : int [1:99] 1 0 0 0 3 5 4 3 11 12 ...
  ..$ density : num [1:99] 0.00153 0 0 0 0.0046 ...
  ..$ mids    : num [1:99] 0.555 0.576 0.597 0.618 0.639 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ apob      :List of 6
  ..$ breaks  : num [1:100] 0.094 0.126 0.159 0.192 0.224 ...
  ..$ counts  : int [1:99] 1 0 0 0 0 0 1 8 17 45 ...
  ..$ density : num [1:99] 0.000985 0 0 0 0 ...
  ..$ mids    : num [1:99] 0.11 0.143 0.175 0.208 0.24 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ totfa     :List of 6
  ..$ breaks  : num [1:100] 3.69 4.09 4.5 4.91 5.32 ...
  ..$ counts  : int [1:99] 7 21 60 97 153 253 392 489 650 779 ...
  ..$ density : num [1:99] 0.000554 0.001662 0.004749 0.007677 0.012109 ...
  ..$ mids    : num [1:99] 3.89 4.3 4.71 5.11 5.52 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ dha       :List of 6
  ..$ breaks  : num [1:100] 0 0.0106 0.0213 0.0319 0.0425 ...
  ..$ counts  : int [1:99] 308 93 151 388 786 1171 1481 1922 2247 2452 ...
  ..$ density : num [1:99] 0.935 0.282 0.459 1.178 2.387 ...
  ..$ mids    : num [1:99] 0.00532 0.01595 0.02659 0.03723 0.04786 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ la        :List of 6
  ..$ breaks  : num [1:100] 0.0408 0.1564 0.2721 0.3877 0.5034 ...
  ..$ counts  : int [1:99] 1 0 5 12 41 62 71 82 106 124 ...
  ..$ density : num [1:99] 0.000279 0 0.001397 0.003352 0.011454 ...
  ..$ mids    : num [1:99] 0.0986 0.2143 0.3299 0.4456 0.5612 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ faw3      :List of 6
  ..$ breaks  : num [1:100] 2.38e-10 2.62e-02 5.24e-02 7.87e-02 1.05e-01 ...
  ..$ counts  : int [1:99] 95 88 132 296 578 737 761 840 1039 1298 ...
  ..$ density : num [1:99] 0.117 0.108 0.163 0.365 0.712 ...
  ..$ mids    : num [1:99] 0.0131 0.0393 0.0656 0.0918 0.118 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ faw6      :List of 6
  ..$ breaks  : num [1:100] 0.85 0.969 1.088 1.207 1.326 ...
  ..$ counts  : int [1:99] 3 19 36 66 90 117 168 191 261 346 ...
  ..$ density : num [1:99] 0.000815 0.005159 0.009775 0.017922 0.024439 ...
  ..$ mids    : num [1:99] 0.909 1.028 1.147 1.266 1.385 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ pufa      :List of 6
  ..$ breaks  : num [1:100] 0.864 0.993 1.123 1.252 1.382 ...
  ..$ counts  : int [1:99] 5 23 34 62 88 114 141 197 224 319 ...
  ..$ density : num [1:99] 0.00125 0.00574 0.00849 0.01549 0.02198 ...
  ..$ mids    : num [1:99] 0.929 1.058 1.187 1.317 1.446 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ mufa      :List of 6
  ..$ breaks  : num [1:100] 0.719 0.886 1.053 1.221 1.388 ...
  ..$ counts  : int [1:99] 9 38 152 388 715 1168 1523 1915 2313 2320 ...
  ..$ density : num [1:99] 0.00174 0.00735 0.0294 0.07505 0.1383 ...
  ..$ mids    : num [1:99] 0.803 0.97 1.137 1.304 1.471 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ sfa       :List of 6
  ..$ breaks  : num [1:100] 1.51 1.66 1.81 1.96 2.11 ...
  ..$ counts  : int [1:99] 5 6 18 39 98 147 253 443 645 870 ...
  ..$ density : num [1:99] 0.00108 0.0013 0.0039 0.00844 0.02121 ...
  ..$ mids    : num [1:99] 1.59 1.74 1.89 2.04 2.18 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ faw3_fa   :List of 6
  ..$ breaks  : num [1:100] 2.88e-09 1.77e-01 3.54e-01 5.30e-01 7.07e-01 ...
  ..$ counts  : int [1:99] 54 26 26 41 49 51 59 127 216 457 ...
  ..$ density : num [1:99] 0.00987 0.00475 0.00475 0.00749 0.00895 ...
  ..$ mids    : num [1:99] 0.0884 0.2652 0.4419 0.6187 0.7955 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ faw6_fa   :List of 6
  ..$ breaks  : num [1:100] 10.1 10.4 10.8 11.2 11.5 ...
  ..$ counts  : int [1:99] 1 0 0 0 0 0 0 0 0 0 ...
  ..$ density : num [1:99] 8.86e-05 0.00 0.00 0.00 0.00 ...
  ..$ mids    : num [1:99] 10.3 10.6 11 11.3 11.7 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ pufa_fa   :List of 6
  ..$ breaks  : num [1:100] 15 15.5 16 16.5 17 ...
  ..$ counts  : int [1:99] 2 1 1 1 2 1 2 6 7 10 ...
  ..$ density : num [1:99] 1.28e-04 6.42e-05 6.42e-05 6.42e-05 1.28e-04 ...
  ..$ mids    : num [1:99] 15.3 15.8 16.3 16.8 17.3 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ mufa_fa   :List of 6
  ..$ breaks  : num [1:100] 8.38 8.77 9.17 9.57 9.97 ...
  ..$ counts  : int [1:99] 1 0 0 1 0 0 0 0 0 0 ...
  ..$ density : num [1:99] 8.15e-05 0.00 0.00 8.15e-05 0.00 ...
  ..$ mids    : num [1:99] 8.58 8.97 9.37 9.77 10.16 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ sfa_fa    :List of 6
  ..$ breaks  : num [1:100] 25.8 26 26.3 26.6 26.9 ...
  ..$ counts  : int [1:99] 1 0 1 0 0 0 0 0 0 1 ...
  ..$ density : num [1:99] 0.000117 0 0.000117 0 0 ...
  ..$ mids    : num [1:99] 25.9 26.2 26.5 26.7 27 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ unsatdeg  :List of 6
  ..$ breaks  : num [1:100] 0.645 0.658 0.672 0.685 0.698 ...
  ..$ counts  : int [1:99] 1 0 1 0 0 1 1 2 1 4 ...
  ..$ density : num [1:99] 0.0024 0 0.0024 0 0 ...
  ..$ mids    : num [1:99] 0.651 0.665 0.678 0.692 0.705 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ xxl_vldl_l:List of 6
  ..$ breaks  : num [1:100] 0 0.00795 0.0159 0.02385 0.0318 ...
  ..$ counts  : int [1:99] 4020 4941 5370 4292 3172 2330 1687 1205 992 702 ...
  ..$ density : num [1:99] 16.3 20 21.7 17.4 12.8 ...
  ..$ mids    : num [1:99] 0.00398 0.01193 0.01988 0.02783 0.03578 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
}
\examples{
data(BBMRI_hist)
## maybe str(BBMRI_hist) ; plot(BBMRI_hist) ...
}
\keyword{datasets}
