% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{BBMRI_hist_plot}
\alias{BBMRI_hist_plot}
\title{Function to plot the 57 selected metabolites to their distributions in BBMRI-nl}
\usage{
BBMRI_hist_plot(
  dat,
  x_name,
  color = c21,
  scaled = FALSE,
  datatype = "metabolite",
  main = "Comparison with the metabolites measures in BBMRI"
)
}
\arguments{
\item{dat}{data.frame or matrix with the metabolites}

\item{x_name}{string with the name of the selected variable}

\item{color}{colors selected for all the variables}

\item{scaled}{TRUE/FALSE. If TRUE the variables will be z-scaled before plotting them}

\item{datatype}{a character vector indicating what data type is beeing plotted}

\item{main}{title of the plot}
}
\value{
plotly image with the histogram of the selected variable compared to the distributions in BBMRI-nl
}
\description{
Function to plot the 57 selected metabolites to their distributions in BBMRI-nl
}
