\name{BBMRI_hist_scaled}
\alias{BBMRI_hist_scaled}
\docType{data}
\title{
Histograms of the metabolites scaled in BBMRI-nl
}
\description{
This list collects the histograms of the 57 metabolites scaled collected in BBMRI-nl with which the user will be able to compare his/her own dataset.
}
\usage{data("BBMRI_hist_scaled")}
\format{
  The format is:
List of 57
 $ ala       :List of 6
  ..$ breaks  : num [1:100] -2.98 -2.89 -2.8 -2.71 -2.62 ...
  ..$ counts  : int [1:99] 1 0 1 2 1 4 11 18 46 49 ...
  ..$ density : num [1:99] 0.000356 0 0.000356 0.000712 0.000356 ...
  ..$ mids    : num [1:99] -2.93 -2.84 -2.75 -2.66 -2.57 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ gln       :List of 6
  ..$ breaks  : num [1:100] -5.32 -5.21 -5.09 -4.98 -4.86 ...
  ..$ counts  : int [1:99] 2 0 1 0 0 1 2 0 1 3 ...
  ..$ density : num [1:99] 0.000647 0 0.000323 0 0 ...
  ..$ mids    : num [1:99] -5.27 -5.15 -5.04 -4.92 -4.8 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ his       :List of 6
  ..$ breaks  : num [1:100] -3.72 -3.3 -2.87 -2.45 -2.03 ...
  ..$ counts  : int [1:99] 7 16 84 299 855 1907 3387 4752 5676 5354 ...
  ..$ density : num [1:99] 0.000533 0.001217 0.006391 0.022748 0.065048 ...
  ..$ mids    : num [1:99] -3.51 -3.08 -2.66 -2.24 -1.82 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ phe       :List of 6
  ..$ breaks  : num [1:100] -2.74 -2.53 -2.32 -2.1 -1.89 ...
  ..$ counts  : int [1:99] 1 3 5 45 153 455 940 1814 2714 3292 ...
  ..$ density : num [1:99] 0.000151 0.000454 0.000756 0.006804 0.023134 ...
  ..$ mids    : num [1:99] -2.64 -2.42 -2.21 -2 -1.79 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ tyr       :List of 6
  ..$ breaks  : num [1:100] -2.84 -2.74 -2.64 -2.54 -2.45 ...
  ..$ counts  : int [1:99] 1 3 7 8 15 28 40 46 80 113 ...
  ..$ density : num [1:99] 0.000326 0.000978 0.002282 0.002608 0.004891 ...
  ..$ mids    : num [1:99] -2.79 -2.69 -2.59 -2.5 -2.4 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ ile       :List of 6
  ..$ breaks  : num [1:100] -2.64 -2.53 -2.43 -2.32 -2.21 ...
  ..$ counts  : int [1:99] 1 0 0 3 9 19 35 58 105 168 ...
  ..$ density : num [1:99] 0.000299 0 0 0.000898 0.002695 ...
  ..$ mids    : num [1:99] -2.59 -2.48 -2.37 -2.27 -2.16 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ leu       :List of 6
  ..$ breaks  : num [1:100] -3.82 -3.69 -3.56 -3.43 -3.29 ...
  ..$ counts  : int [1:99] 1 0 0 0 1 0 1 1 4 6 ...
  ..$ density : num [1:99] 0.000241 0 0 0 0.000241 ...
  ..$ mids    : num [1:99] -3.76 -3.63 -3.49 -3.36 -3.23 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ val       :List of 6
  ..$ breaks  : num [1:100] -3.94 -3.82 -3.7 -3.57 -3.45 ...
  ..$ counts  : int [1:99] 2 2 3 2 3 1 5 5 6 9 ...
  ..$ density : num [1:99] 0.000514 0.000514 0.000771 0.000514 0.000771 ...
  ..$ mids    : num [1:99] -3.88 -3.76 -3.63 -3.51 -3.38 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ glc       :List of 6
  ..$ breaks  : num [1:100] -2.9 -2.76 -2.61 -2.47 -2.33 ...
  ..$ counts  : int [1:99] 88 50 56 56 69 69 92 98 108 143 ...
  ..$ density : num [1:99] 0.02 0.0113 0.0127 0.0127 0.0157 ...
  ..$ mids    : num [1:99] -2.83 -2.68 -2.54 -2.4 -2.26 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ lac       :List of 6
  ..$ breaks  : num [1:100] -1.05 -0.912 -0.774 -0.636 -0.498 ...
  ..$ counts  : int [1:99] 85 361 1837 4156 5210 4875 3994 2944 1972 1342 ...
  ..$ density : num [1:99] 0.0197 0.0838 0.4265 0.965 1.2097 ...
  ..$ mids    : num [1:99] -0.981 -0.843 -0.705 -0.567 -0.429 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ cit       :List of 6
  ..$ breaks  : num [1:100] -0.566 0.51 1.586 2.662 3.738 ...
  ..$ counts  : int [1:99] 30968 222 2 0 0 0 1 1 0 0 ...
  ..$ density : num [1:99] 9.22e-01 6.61e-03 5.96e-05 0.00 0.00 ...
  ..$ mids    : num [1:99] -0.0282 1.048 2.1242 3.2004 4.2766 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ ace       :List of 6
  ..$ breaks  : num [1:100] -1.197 -0.693 -0.188 0.317 0.822 ...
  ..$ counts  : int [1:99] 1542 13598 9185 4641 1479 360 107 38 20 12 ...
  ..$ density : num [1:99] 0.0981 0.8652 0.5844 0.2953 0.0941 ...
  ..$ mids    : num [1:99] -0.9451 -0.4403 0.0646 0.5694 1.0742 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ acace     :List of 6
  ..$ breaks  : num [1:100] -1.236 -0.91 -0.584 -0.257 0.069 ...
  ..$ counts  : int [1:99] 1213 5277 8407 6427 3659 2176 1201 710 502 337 ...
  ..$ density : num [1:99] 0.12 0.522 0.831 0.635 0.362 ...
  ..$ mids    : num [1:99] -1.0728 -0.7466 -0.4204 -0.0942 0.2321 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ crea      :List of 6
  ..$ breaks  : num [1:100] -2.397 -1.996 -1.595 -1.195 -0.794 ...
  ..$ counts  : int [1:99] 3 24 460 3132 7060 7698 5700 3338 1650 831 ...
  ..$ density : num [1:99] 0.000241 0.001929 0.036976 0.251758 0.567501 ...
  ..$ mids    : num [1:99] -2.196 -1.796 -1.395 -0.994 -0.594 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ alb       :List of 6
  ..$ breaks  : num [1:100] -10.62 -10.43 -10.25 -10.06 -9.88 ...
  ..$ counts  : int [1:99] 1 0 0 0 0 0 0 0 0 0 ...
  ..$ density : num [1:99] 0.000174 0 0 0 0 ...
  ..$ mids    : num [1:99] -10.52 -10.34 -10.16 -9.97 -9.79 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ gp        :List of 6
  ..$ breaks  : num [1:100] -3.28 -3.13 -2.98 -2.83 -2.68 ...
  ..$ counts  : int [1:99] 2 2 1 2 3 11 27 37 58 159 ...
  ..$ density : num [1:99] 0.000424 0.000424 0.000212 0.000424 0.000636 ...
  ..$ mids    : num [1:99] -3.21 -3.06 -2.91 -2.75 -2.6 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ m_vldl_l  :List of 6
  ..$ breaks  : num [1:100] -1.625 -1.465 -1.306 -1.146 -0.987 ...
  ..$ counts  : int [1:99] 34 127 655 1602 2494 3010 3000 2894 2627 2267 ...
  ..$ density : num [1:99] 0.00683 0.0255 0.13149 0.3216 0.50067 ...
  ..$ mids    : num [1:99] -1.545 -1.386 -1.226 -1.066 -0.907 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ s_vldl_l  :List of 6
  ..$ breaks  : num [1:100] -2.78 -2.62 -2.47 -2.32 -2.17 ...
  ..$ counts  : int [1:99] 19 0 1 6 28 92 217 346 585 790 ...
  ..$ density : num [1:99] 0.00398 0 0.00021 0.00126 0.00587 ...
  ..$ mids    : num [1:99] -2.7 -2.55 -2.39 -2.24 -2.09 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ xs_vldl_l :List of 6
  ..$ breaks  : num [1:100] -4.05 -3.92 -3.79 -3.66 -3.53 ...
  ..$ counts  : int [1:99] 10 0 0 0 0 0 0 6 9 19 ...
  ..$ density : num [1:99] 0.00245 0 0 0 0 ...
  ..$ mids    : num [1:99] -3.99 -3.85 -3.72 -3.59 -3.46 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ idl_l     :List of 6
  ..$ breaks  : num [1:100] -3.42 -3.28 -3.15 -3.01 -2.87 ...
  ..$ counts  : int [1:99] 1 0 4 4 16 34 33 62 58 99 ...
  ..$ density : num [1:99] 0.000236 0 0.000943 0.000943 0.003774 ...
  ..$ mids    : num [1:99] -3.35 -3.21 -3.08 -2.94 -2.81 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ l_ldl_l   :List of 6
  ..$ breaks  : num [1:100] -3.12 -2.99 -2.85 -2.72 -2.58 ...
  ..$ counts  : int [1:99] 32 0 1 2 18 30 58 119 140 261 ...
  ..$ density : num [1:99] 0.00755 0 0.000236 0.000472 0.004247 ...
  ..$ mids    : num [1:99] -3.06 -2.92 -2.79 -2.65 -2.51 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ m_ldl_l   :List of 6
  ..$ breaks  : num [1:100] -2.89 -2.76 -2.63 -2.5 -2.37 ...
  ..$ counts  : int [1:99] 134 0 0 0 3 46 101 151 260 373 ...
  ..$ density : num [1:99] 0.033164 0 0 0 0.000742 ...
  ..$ mids    : num [1:99] -2.82 -2.69 -2.56 -2.43 -2.3 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ s_ldl_l   :List of 6
  ..$ breaks  : num [1:100] -3.1 -2.97 -2.84 -2.71 -2.58 ...
  ..$ counts  : int [1:99] 134 0 0 0 3 8 42 69 135 201 ...
  ..$ density : num [1:99] 0.032903 0 0 0 0.000737 ...
  ..$ mids    : num [1:99] -3.04 -2.91 -2.78 -2.65 -2.52 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ m_hdl_l   :List of 6
  ..$ breaks  : num [1:100] -3.43 -3.35 -3.27 -3.19 -3.1 ...
  ..$ counts  : int [1:99] 345 0 0 0 0 0 1 0 0 0 ...
  ..$ density : num [1:99] 0.134 0 0 0 0 ...
  ..$ mids    : num [1:99] -3.39 -3.31 -3.23 -3.14 -3.06 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ s_hdl_l   :List of 6
  ..$ breaks  : num [1:100] -6.64 -6.52 -6.4 -6.28 -6.16 ...
  ..$ counts  : int [1:99] 39 0 0 0 0 0 0 0 0 0 ...
  ..$ density : num [1:99] 0.0105 0 0 0 0 ...
  ..$ mids    : num [1:99] -6.58 -6.46 -6.34 -6.22 -6.1 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ idl_c     :List of 6
  ..$ breaks  : num [1:100] -3.33 -3.19 -3.06 -2.92 -2.79 ...
  ..$ counts  : int [1:99] 33 3 8 7 25 33 51 58 81 127 ...
  ..$ density : num [1:99] 0.00781 0.00071 0.00189 0.00166 0.00592 ...
  ..$ mids    : num [1:99] -3.26 -3.13 -2.99 -2.85 -2.72 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ serum_c   :List of 6
  ..$ breaks  : num [1:100] -3.65 -3.52 -3.38 -3.25 -3.11 ...
  ..$ counts  : int [1:99] 1 0 2 8 19 20 39 38 50 66 ...
  ..$ density : num [1:99] 0.000238 0 0.000476 0.001905 0.004524 ...
  ..$ mids    : num [1:99] -3.59 -3.45 -3.32 -3.18 -3.05 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ vldl_c    :List of 6
  ..$ breaks  : num [1:100] -2.77 -2.61 -2.46 -2.3 -2.15 ...
  ..$ counts  : int [1:99] 1 4 18 47 111 169 291 399 558 772 ...
  ..$ density : num [1:99] 0.000208 0.000832 0.003743 0.009773 0.023081 ...
  ..$ mids    : num [1:99] -2.69 -2.54 -2.38 -2.23 -2.07 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ ldl_c     :List of 6
  ..$ breaks  : num [1:100] -2.71 -2.58 -2.45 -2.32 -2.19 ...
  ..$ counts  : int [1:99] 8 23 29 56 104 141 237 301 413 598 ...
  ..$ density : num [1:99] 0.00198 0.0057 0.00719 0.01389 0.02579 ...
  ..$ mids    : num [1:99] -2.64 -2.51 -2.38 -2.25 -2.12 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ hdl_c     :List of 6
  ..$ breaks  : num [1:100] -3.53 -3.44 -3.36 -3.27 -3.18 ...
  ..$ counts  : int [1:99] 1 0 1 1 1 6 3 5 10 16 ...
  ..$ density : num [1:99] 0.000363 0 0.000363 0.000363 0.000363 ...
  ..$ mids    : num [1:99] -3.49 -3.4 -3.31 -3.22 -3.13 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ hdl2_c    :List of 6
  ..$ breaks  : num [1:100] -2.59 -2.51 -2.43 -2.35 -2.27 ...
  ..$ counts  : int [1:99] 124 32 47 53 46 80 82 83 101 128 ...
  ..$ density : num [1:99] 0.0494 0.0127 0.0187 0.0211 0.0183 ...
  ..$ mids    : num [1:99] -2.55 -2.47 -2.39 -2.31 -2.23 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ hdl3_c    :List of 6
  ..$ breaks  : num [1:100] -6.42 -6.27 -6.11 -5.96 -5.8 ...
  ..$ counts  : int [1:99] 1 0 0 0 0 0 1 0 0 0 ...
  ..$ density : num [1:99] 0.000207 0 0 0 0 ...
  ..$ mids    : num [1:99] -6.34 -6.19 -6.03 -5.88 -5.73 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ vldl_d    :List of 6
  ..$ breaks  : num [1:100] -3.03 -2.95 -2.86 -2.78 -2.69 ...
  ..$ counts  : int [1:99] 1 0 1 1 3 1 5 13 13 14 ...
  ..$ density : num [1:99] 0.00038 0 0.00038 0.00038 0.00114 ...
  ..$ mids    : num [1:99] -2.99 -2.9 -2.82 -2.74 -2.65 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ ldl_d     :List of 6
  ..$ breaks  : num [1:100] -11.4 -11.2 -11 -10.7 -10.5 ...
  ..$ counts  : int [1:99] 2 1 1 0 0 0 0 2 0 0 ...
  ..$ density : num [1:99] 0.000267 0.000133 0.000133 0 0 ...
  ..$ mids    : num [1:99] -11.3 -11.1 -10.8 -10.6 -10.4 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ hdl_d     :List of 6
  ..$ breaks  : num [1:100] -3.72 -3.64 -3.56 -3.49 -3.41 ...
  ..$ counts  : int [1:99] 1 0 0 0 0 1 3 2 1 3 ...
  ..$ density : num [1:99] 0.000413 0 0 0 0 ...
  ..$ mids    : num [1:99] -3.68 -3.6 -3.53 -3.45 -3.37 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ serum_tg  :List of 6
  ..$ breaks  : num [1:100] -1.77 -1.6 -1.43 -1.26 -1.09 ...
  ..$ counts  : int [1:99] 6 61 308 944 1926 2894 3040 3214 2892 2666 ...
  ..$ density : num [1:99] 0.00114 0.0116 0.05858 0.17956 0.36634 ...
  ..$ mids    : num [1:99] -1.68 -1.52 -1.35 -1.18 -1.01 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ totpg     :List of 6
  ..$ breaks  : num [1:100] -3.75 -3.65 -3.55 -3.45 -3.35 ...
  ..$ counts  : int [1:99] 2 0 2 3 5 2 8 30 25 35 ...
  ..$ density : num [1:99] 0.000641 0 0.000641 0.000962 0.001603 ...
  ..$ mids    : num [1:99] -3.7 -3.6 -3.5 -3.4 -3.3 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ pc        :List of 6
  ..$ breaks  : num [1:100] -4.53 -4.41 -4.29 -4.17 -4.06 ...
  ..$ counts  : int [1:99] 14 0 0 0 1 0 4 6 4 25 ...
  ..$ density : num [1:99] 0.003872 0 0 0 0.000277 ...
  ..$ mids    : num [1:99] -4.47 -4.35 -4.23 -4.12 -4 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ sm        :List of 6
  ..$ breaks  : num [1:100] -3.5 -3.39 -3.28 -3.17 -3.06 ...
  ..$ counts  : int [1:99] 1 1 0 5 10 8 13 25 30 42 ...
  ..$ density : num [1:99] 0.000293 0.000293 0 0.001465 0.002931 ...
  ..$ mids    : num [1:99] -3.45 -3.34 -3.23 -3.12 -3.01 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ totcho    :List of 6
  ..$ breaks  : num [1:100] -4.89 -4.78 -4.66 -4.55 -4.43 ...
  ..$ counts  : int [1:99] 1 0 0 0 0 0 0 0 2 2 ...
  ..$ density : num [1:99] 0.000279 0 0 0 0 ...
  ..$ mids    : num [1:99] -4.84 -4.72 -4.61 -4.49 -4.38 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ apoa1     :List of 6
  ..$ breaks  : num [1:100] -4.24 -4.15 -4.06 -3.97 -3.88 ...
  ..$ counts  : int [1:99] 1 0 0 0 3 5 4 3 11 12 ...
  ..$ density : num [1:99] 0.000358 0 0 0 0.001073 ...
  ..$ mids    : num [1:99] -4.19 -4.1 -4.01 -3.93 -3.84 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ apob      :List of 6
  ..$ breaks  : num [1:100] -3.71 -3.56 -3.42 -3.28 -3.14 ...
  ..$ counts  : int [1:99] 1 0 0 0 0 0 1 8 17 45 ...
  ..$ density : num [1:99] 0.000225 0 0 0 0 ...
  ..$ mids    : num [1:99] -3.64 -3.49 -3.35 -3.21 -3.07 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ totfa     :List of 6
  ..$ breaks  : num [1:100] -2.69 -2.54 -2.39 -2.25 -2.1 ...
  ..$ counts  : int [1:99] 7 21 60 97 153 253 392 489 650 779 ...
  ..$ density : num [1:99] 0.00154 0.00461 0.01316 0.02127 0.03355 ...
  ..$ mids    : num [1:99] -2.62 -2.47 -2.32 -2.17 -2.03 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ dha       :List of 6
  ..$ breaks  : num [1:100] -2.21 -2.03 -1.85 -1.67 -1.49 ...
  ..$ counts  : int [1:99] 308 93 151 388 786 1171 1481 1922 2247 2452 ...
  ..$ density : num [1:99] 0.0557 0.0168 0.0273 0.0701 0.1421 ...
  ..$ mids    : num [1:99] -2.12 -1.94 -1.76 -1.58 -1.4 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ la        :List of 6
  ..$ breaks  : num [1:100] -3.58 -3.43 -3.29 -3.14 -3 ...
  ..$ counts  : int [1:99] 1 0 5 12 41 62 71 82 106 124 ...
  ..$ density : num [1:99] 0.000223 0 0.001114 0.002674 0.009135 ...
  ..$ mids    : num [1:99] -3.51 -3.36 -3.22 -3.07 -2.93 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ faw3      :List of 6
  ..$ breaks  : num [1:100] -2.45 -2.28 -2.12 -1.95 -1.78 ...
  ..$ counts  : int [1:99] 95 88 132 296 578 737 761 840 1039 1298 ...
  ..$ density : num [1:99] 0.0184 0.017 0.0255 0.0573 0.1119 ...
  ..$ mids    : num [1:99] -2.37 -2.2 -2.03 -1.87 -1.7 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ faw6      :List of 6
  ..$ breaks  : num [1:100] -3.09 -2.96 -2.82 -2.69 -2.56 ...
  ..$ counts  : int [1:99] 3 19 36 66 90 117 168 191 261 346 ...
  ..$ density : num [1:99] 0.000736 0.004658 0.008826 0.016181 0.022065 ...
  ..$ mids    : num [1:99] -3.02 -2.89 -2.76 -2.63 -2.49 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ pufa      :List of 6
  ..$ breaks  : num [1:100] -3.12 -2.99 -2.86 -2.74 -2.61 ...
  ..$ counts  : int [1:99] 5 23 34 62 88 114 141 197 224 319 ...
  ..$ density : num [1:99] 0.00126 0.00582 0.0086 0.01568 0.02226 ...
  ..$ mids    : num [1:99] -3.05 -2.93 -2.8 -2.67 -2.54 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ mufa      :List of 6
  ..$ breaks  : num [1:100] -2.2 -2.03 -1.86 -1.68 -1.51 ...
  ..$ counts  : int [1:99] 9 38 152 388 715 1168 1523 1915 2313 2320 ...
  ..$ density : num [1:99] 0.0017 0.00717 0.02869 0.07324 0.13496 ...
  ..$ mids    : num [1:99] -2.11 -1.94 -1.77 -1.6 -1.43 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ sfa       :List of 6
  ..$ breaks  : num [1:100] -2.69 -2.54 -2.39 -2.25 -2.1 ...
  ..$ counts  : int [1:99] 5 6 18 39 98 147 253 443 645 870 ...
  ..$ density : num [1:99] 0.0011 0.00132 0.00397 0.0086 0.02161 ...
  ..$ mids    : num [1:99] -2.61 -2.47 -2.32 -2.17 -2.03 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ faw3_fa   :List of 6
  ..$ breaks  : num [1:100] -3.17 -3.01 -2.84 -2.68 -2.51 ...
  ..$ counts  : int [1:99] 54 26 26 41 49 51 59 127 216 457 ...
  ..$ density : num [1:99] 0.0106 0.0051 0.0051 0.00805 0.00962 ...
  ..$ mids    : num [1:99] -3.09 -2.92 -2.76 -2.6 -2.43 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ faw6_fa   :List of 6
  ..$ breaks  : num [1:100] -5.95 -5.85 -5.75 -5.66 -5.56 ...
  ..$ counts  : int [1:99] 1 0 0 0 0 0 0 0 0 0 ...
  ..$ density : num [1:99] 0.000339 0 0 0 0 ...
  ..$ mids    : num [1:99] -5.9 -5.8 -5.71 -5.61 -5.52 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ pufa_fa   :List of 6
  ..$ breaks  : num [1:100] -5.1 -4.98 -4.86 -4.74 -4.61 ...
  ..$ counts  : int [1:99] 2 1 1 1 2 1 2 6 7 10 ...
  ..$ density : num [1:99] 0.000533 0.000266 0.000266 0.000266 0.000533 ...
  ..$ mids    : num [1:99] -5.04 -4.92 -4.8 -4.68 -4.55 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ mufa_fa   :List of 6
  ..$ breaks  : num [1:100] -4.59 -4.49 -4.38 -4.27 -4.16 ...
  ..$ counts  : int [1:99] 1 0 0 1 0 0 0 0 0 0 ...
  ..$ density : num [1:99] 0.000302 0 0 0.000302 0 ...
  ..$ mids    : num [1:99] -4.54 -4.43 -4.32 -4.22 -4.11 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ sfa_fa    :List of 6
  ..$ breaks  : num [1:100] -4.68 -4.58 -4.48 -4.37 -4.27 ...
  ..$ counts  : int [1:99] 1 0 1 0 0 0 0 0 0 1 ...
  ..$ density : num [1:99] 0.000314 0 0.000314 0 0 ...
  ..$ mids    : num [1:99] -4.63 -4.53 -4.43 -4.32 -4.22 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ unsatdeg  :List of 6
  ..$ breaks  : num [1:100] -5.13 -5 -4.87 -4.74 -4.61 ...
  ..$ counts  : int [1:99] 1 0 1 0 0 1 1 2 1 4 ...
  ..$ density : num [1:99] 0.000252 0 0.000252 0 0 ...
  ..$ mids    : num [1:99] -5.06 -4.94 -4.81 -4.68 -4.55 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
 $ xxl_vldl_l:List of 6
  ..$ breaks  : num [1:100] -1.0243 -0.7878 -0.5513 -0.3147 -0.0782 ...
  ..$ counts  : int [1:99] 4020 4941 5370 4292 3172 2330 1687 1205 992 702 ...
  ..$ density : num [1:99] 0.547 0.673 0.731 0.584 0.432 ...
  ..$ mids    : num [1:99] -0.906 -0.67 -0.433 -0.196 0.04 ...
  ..$ xname   : chr "na.omit(x)"
  ..$ equidist: logi TRUE
  ..- attr(*, "class")= chr "histogram"
}
\examples{
data(BBMRI_hist_scaled)
## maybe str(BBMRI_hist_scaled) ; plot(BBMRI_hist_scaled) ...
}
\keyword{datasets}
