% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{binarize_all_pheno}
\alias{binarize_all_pheno}
\title{Function created to binarize all the phenotypes supplied using the threshold defined
in the surrogates by Bizzarri et al.}
\usage{
binarize_all_pheno(data)
}
\arguments{
\item{data}{phenotypes data.frame containing some of the following variables (with the same namings):
"sex","diabetes", "lipidmed",  "blood_pressure_lowering_med", "current_smoking",
"metabolic_syndrome", "alcohol_consumption", "age","BMI", "ln_hscrp","waist_circumference",
"weight","height", "triglycerides", "ldl_chol", "hdlchol", "totchol", "eGFR","wbc","hgb"}
}
\value{
The phenotypic variables binarized following the thresholds in in the surrogates by Bizzarri et al.
}
\description{
Function created to binarize all the phenotypes supplied using the threshold defined
in the surrogates by Bizzarri et al.
}
