% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{BMI_LDL_eGFR}
\alias{BMI_LDL_eGFR}
\title{Function created to calculate:
the BMI using height and weight;
the LDL cholesterol using HDL cholesterol, triglycerides, totchol;
eGFR using creatinine, sex and age.}
\usage{
BMI_LDL_eGFR(phenotypes, metabo_measures)
}
\arguments{
\item{phenotypes}{data.frame containing height and weight, HDL cholesterol, triglycerides, totchol, sex and age}

\item{metabo_measures}{NH-metabolomics matrix containing creatinine (crea)}
}
\value{
The phenotypes data.frame with the addition of BMI, LDL cholesterol and eGFR
}
\description{
Function created to calculate:
the BMI using height and weight;
the LDL cholesterol using HDL cholesterol, triglycerides, totchol;
eGFR using creatinine, sex and age.
}
