% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{c21}
\alias{c21}
\title{c21}
\format{
An object of class \code{character} of length 21.
}
\usage{
data("c21")
}
\description{
Colors attributed to each metabolomics-based model in MiMIR
}
\examples{
data("c21")

}
\keyword{datasets}
