% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{calculate_surrogate_scores}
\alias{calculate_surrogate_scores}
\title{Function created to calculate the surrogate scores by Bizzarri et al. from  the NH-metabolomics matrix}
\usage{
calculate_surrogate_scores(
  met,
  pheno,
  PARAM_surrogates,
  bin_names,
  Nmax_miss = 1,
  Nmax_zero = 1,
  post = TRUE,
  roc = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{met}{The NH-metabolomics matrix}

\item{pheno}{The phenotypic matrix including this clinical variables: "sex","diabetes", "lipidmed",  "blood_pressure_lowering_med", "current_smoking",
"metabolic_syndrome", "alcohol_consumption", "age","BMI", "ln_hscrp","waist_circumference",
"weight","height", "triglycerides", "ldl_chol", "hdlchol", "totchol", "eGFR","wbc","hgb"}

\item{PARAM_surrogates}{is a list holding the parameters to compute the surrogates}

\item{bin_names}{string vector indicating the names of the binary variables obtained}

\item{Nmax_miss}{max number of missing values allowed per sample}

\item{Nmax_zero}{max number of zeros allowed per sample}

\item{post}{TRUE/FALSE. If TRUE it will calculate the posterior probabilities}

\item{roc}{TRUE/FALSE. If TRUE it will produce an image with all the roc curves
for the surrogate with the avialable phenotypes in the current dataset}

\item{quiet}{TRUE/FALSE. if TRUE if will suppress all messages from the function}
}
\value{
if pheno is not available: a list with the surrogates and the NH metabolomics matrix after QC.
if pheno is available: a list with the surrogates, the roc curves, the phenotypes, the binarized phenotypes and the NH metabolomics matrix after QC,
}
\description{
Function created to calculate the surrogate scores by Bizzarri et al. from  the NH-metabolomics matrix
}
