% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{calib_data_frame}
\alias{calib_data_frame}
\title{helper function that creates a data.frame with the Platt Calibrations}
\usage{
calib_data_frame(calibrations, bin_phenotypes, bin_pheno_available)
}
\arguments{
\item{calibrations}{list result of calibration_surro}

\item{bin_phenotypes}{data.frame with binary phenotypes resulting form binarize_all_pheno}

\item{bin_pheno_available}{string with the names of the binarize clinical variables available in the dataset}
}
\value{
data.frame with the calibrated surrogates
}
\description{
helper function that creates a data.frame with the Platt Calibrations
}
