% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{calibration_surro}
\alias{calibration_surro}
\title{helper function that calculates the Platt Calibrations}
\usage{
calibration_surro(
  bin_phenotypes,
  surrogates,
  bin_names,
  bin_pheno_available,
  pl = FALSE,
  nbins = 10
)
}
\arguments{
\item{bin_phenotypes}{data.frame with binary phenotypes resulting form binarize_all_pheno}

\item{surrogates}{data.frame with surrogates resulting from calculate_surrogate_scores}

\item{bin_names}{string with the names of the binarize clinical variables}

\item{bin_pheno_available}{string with the names of the binarize clinical variables available in the dataset}

\item{pl}{TRUE/FALSE. If TRUE creates the calibration plots}

\item{nbins}{number of bins for the plots}
}
\value{
list with the calibrated surrogates
}
\description{
helper function that calculates the Platt Calibrations
}
