% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{comp.mort_score}
\alias{comp.mort_score}
\title{Function to compute Deelen et al. mortality score on metabolite data.}
\usage{
comp.mort_score(dat, betas = mort_betas, quiet = FALSE)
}
\arguments{
\item{dat}{The NH-metabolomics matrix}

\item{betas}{The betas of the linear regression composing the mortality score}

\item{quiet}{TRUE/FALSE if TRUE if will suppress all messages from the function}
}
\value{
The mortality score by Deelen et al.
}
\description{
Function to compute Deelen et al. mortality score on metabolite data.
}
