% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{comp_covid_score}
\alias{comp_covid_score}
\title{Function to compute COVID score by Nightingale Health from the metabolite data.}
\usage{
comp_covid_score(dat, betas = covid_betas, quiet = FALSE)
}
\arguments{
\item{dat}{The NH-metabolomics matrix}

\item{betas}{The betas of the linear regression composing the score}

\item{quiet}{TRUE/FALSE if TRUE if will suppress all messages from the function}
}
\value{
The COVID score by Nightingale Health
}
\description{
Function to compute COVID score by Nightingale Health from the metabolite data.
}
