% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{cor_assoc}
\alias{cor_assoc}
\title{cor_assoc}
\usage{
cor_assoc(dat1, dat2, feat1, feat2, method = "pearson", quiet = FALSE)
}
\arguments{
\item{dat1}{matrix 1}

\item{dat2}{matrix 2}

\item{feat1}{vector of strings with the names of the selected variables in dat}

\item{feat2}{vector if strings with the names of the selected variables in dat2}

\item{method}{indicates which methods of the correlation to use}

\item{quiet}{logical to suppress the messages in the console}
}
\value{
correlations of the selected variables in the 2 martrices
}
\description{
Function to calulate the correlation between 2 matrices
}
\examples{
library(stats)

#load the dataset
m <- as.matrix(synthetic_metabolic_dataset)

#Compute the pearson correlation of all the variables in the data.frame metabolic_measures
cors<-cor_assoc(m, m, MiMIR::metabolites_subsets$MET63,MiMIR::metabolites_subsets$MET63)

}
\seealso{
plot_corply
}
