% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{covid_betas}
\alias{covid_betas}
\title{COVID-score betas}
\format{
An object of class \code{data.frame} with 25 rows and 3 columns.
}
\usage{
data("covid_betas")
}
\description{
The coefficients used to compute the COVID score by Nightingale Health UK Biobank Initiative et al.
}
\details{
Dataframe containing the abbreviation of the metabolites, the metabolites names and finally the Coefficients to compute the COVID score
}
\examples{
data("covid_betas")

}
\references{
Nightingale Health UK Biobank Initiative et al. (2021) Metabolic biomarker profiling for identification of susceptibility to severe pneumonia and COVID-19 in the general population. eLife, 10, e63033, \url{doi:10.7554/eLife.63033}
}
\keyword{datasets}
