% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CVD_score_betas}
\alias{CVD_score_betas}
\title{CVD-score betas}
\format{
An object of class \code{data.frame} with 12 rows and 3 columns.
}
\usage{
data("CVD_score_betas")
}
\description{
The coefficients used to compute the CVD score by Wurtz et al.
}
\details{
Dataframe containing the abbreviation of the metabolites, the metabolites names and finally the Coefficients to compute the COVID score
}
\examples{
data("CVD_score_betas")

}
\references{
Wurtz,P. et al. (2015) Metabolite profiling and cardiovascular event risk: a prospective study of 3 population-based cohorts. Circulation, 131, 774-785, \url{doi:10.1161/CIRCULATIONAHA.114.013116}
}
\keyword{datasets}
