% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{do.metabowas}
\alias{do.metabowas}
\title{Helper function to do a MetaboWAS}
\usage{
\method{do}{metabowas}(
  phen,
  dat,
  test_variable = "age",
  covariates = c("sex"),
  PC = NULL,
  quiet = T
)
}
\arguments{
\item{phen}{phenotypes data.frame}

\item{dat}{metabolites data.frame}

\item{test_variable}{the variable to be investigated}

\item{covariates}{the covariates that you want to add}

\item{PC}{the number of Principal compontents that represent the 90\% of the variance, in case you want to correct for this number (instead of Bonferroni/BH)}

\item{quiet}{if FALSE it will plot the amount of people avaialble}
}
\value{
results= the results of the MetaboWAS (estimate, tstatistics, pvalue, BH corrected pvalue)
}
\description{
Helper function to do a MetaboWAS
}
