% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{do_metabowas}
\alias{do_metabowas}
\title{Function to compute MetaboWASs}
\usage{
do_metabowas(
  phen,
  dat,
  test_variable = "age",
  covariates = c("sex"),
  adj_method = "BH",
  quiet = T
)
}
\arguments{
\item{phen}{phenotypes data.frame}

\item{dat}{metabolites data.frame}

\item{test_variable}{the variable to be investigated}

\item{covariates}{the covariates that you want to add}

\item{adj_method}{correction method.}

\item{quiet}{if FALSE it will plot the amount of people avaialble}
}
\value{
results= the results of the MetaboWAS (estimate, tstatistics, pvalue, BH corrected pvalue)
}
\description{
This helper function is called when doing Metabolites wide association analysis.It reports the results of linear regression models to study the association of a test variable to each metabolites individually and corrected for the covariates indicated.
}
