% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGenicFeatures.R
\name{extractGenicFeatures}
\alias{extractGenicFeatures}
\title{Generate a list of genic features GRangesList objects}
\usage{
extractGenicFeatures(TxDb = NULL, tx2gene = NULL, selectGn = NULL,
  selectTx = NULL, excludeIntrons = TRUE, bins = c(20, 100, 70, 100),
  verbose = TRUE)
}
\arguments{
\item{TxDb}{A \code{\link[GenomicFeatures:TxDb-class]{GenomicFeatures}} object. It must contain \code{\link[GenomeInfoDb:Seqinfo-class]{GenomeInfoDb}} information.}

\item{tx2gene}{A \code{"\linkS4class{data.frame}"} object. The first column must be of transcript identifiers, while the second must be of gene identifiers. Additional columns will be discarded.}

\item{selectGn}{A vector of optional gene identifiers to keep.}

\item{selectTx}{A vector of optional transcript identifiers to keep.}

\item{excludeIntrons}{When set to 'TRUE', the extraction of intronic regions is skipped.}

\item{bins}{A 4 integers ordered vector. The vector order determines the 5'UTR, CDS, 3'UTR and Introns minumum region widths.}

\item{verbose}{When set to 'TRUE', the function prints diagnostic messages.}
}
\value{
A named list of \code{\link[GenomicRanges:GRangesList-class]{GenomicRanges}} objects.
}
\description{
This function extracts and filters the genic features from a \code{\link[GenomicFeatures:TxDb-class]{GenomicFeatures}} object.
}
\examples{
library("TxDb.Dmelanogaster.UCSC.dm3.ensGene")

genicRegions = extractGenicFeatures(TxDb = TxDb.Dmelanogaster.UCSC.dm3.ensGene)
}
