% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGenomicFeatures.R
\name{extractGenomicFeatures}
\alias{extractGenomicFeatures}
\title{Generate a list of genomic features GRangesList objects}
\usage{
extractGenomicFeatures(TxDb = NULL, selectGn = NULL,
  excludeIntrons = TRUE, body_width = 1000, upstream_width = 100,
  downstream_width = 1000, verbose = TRUE)
}
\arguments{
\item{TxDb}{A \code{\link[GenomicFeatures:TxDb-class]{GenomicFeatures}} object. It must contain \code{\link[GenomeInfoDb:Seqinfo-class]{GenomeInfoDb}} information.}

\item{selectGn}{A vector of optional gene identifiers to keep.}

\item{excludeIntrons}{When set to ‘TRUE’, the extraction of intronic regions is skipped.}

\item{body_width}{A positive integer. It determines the minumum gene body width.}

\item{upstream_width}{A positive integer. It determines the upstream region width.}

\item{downstream_width}{A positive integer. It determines the downstream gene body width.}

\item{verbose}{When set to ‘TRUE’, the function prints diagnostic messages.}
}
\value{
A named list of \code{\link[GenomicRanges:GRangesList-class]{GenomicRanges}} objects.
}
\description{
This function extracts and filters gene, upstream and downstream features from a \code{\link[GenomicFeatures:TxDb-class]{GenomicFeatures}} object.
}
\examples{
library("TxDb.Dmelanogaster.UCSC.dm3.ensGene")

genomicRegions = extractGenomicFeatures(TxDb = TxDb.Dmelanogaster.UCSC.dm3.ensGene)
}
