% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixSeqInfo.R
\name{fixSeqInfo}
\alias{fixSeqInfo}
\title{Generate a list of genomic features GRangesList objects}
\usage{
fixSeqInfo(first = NULL, second = NULL, TxDb = NULL,
  method = c("first", "second", "union", "txdb"))
}
\arguments{
\item{first}{A \code{\link[GenomicRanges:GenomicRanges-class]{GenomicRanges}} or \code{\link[GenomicRanges:GRangesList-class]{GenomicRanges}} object.}

\item{second}{An optional \code{\link[GenomicRanges:GenomicRanges-class]{GenomicRanges}} or \code{\link[GenomicRanges:GRangesList-class]{GenomicRanges}} object.}

\item{TxDb}{A \code{\link[GenomicFeatures:TxDb-class]{GenomicFeatures}} object. It must contain \code{\link[GenomeInfoDb:Seqinfo-class]{GenomeInfoDb}} information.}

\item{method}{A character string indicating which reference should be used to assign the \code{\link{seqlevels}}. One of 'first' (default), 'second', 'union', "txdb": can be abbreviated.}
}
\value{
A list of \code{\link[GenomicRanges:GenomicRanges-class]{GenomicRanges}} or \code{\link[GenomicRanges:GRangesList-class]{GenomicRanges}} object(s).
}
\description{
This function extracts and filters gene, upstream and downstream features from a \code{\link[GenomicFeatures:TxDb-class]{GenomicFeatures}} object.
}
\examples{
library("TxDb.Dmelanogaster.UCSC.dm3.ensGene")

TxDb = TxDb.Dmelanogaster.UCSC.dm3.ensGene

first = GRanges(c("chr4", "chr2L"), IRanges(c(10, 20), c(30, 30)), c("-", "+"))
second = GRanges(c("chr4", "chr3R"), IRanges(c(10, 20), c(30, 30)), c("-", "+"))

# Assign seqlevels and seqinfo from a TxDb object
fixList = fixSeqInfo(first, TxDb=TxDb)
sapply(fixList, head)
sapply(fixList, seqlevels)
sapply(fixList, seqinfo)

# Assign the seqlevels of the first to the second object
fixList = fixSeqInfo(first, second, method="first")
sapply(fixList, head)
sapply(fixList, seqlevels)
sapply(fixList, seqinfo)

# Assign the union of the seqlevels to both objects
fixList = fixSeqInfo(first, second, method="union")
sapply(fixList, head)
sapply(fixList, seqlevels)
sapply(fixList, seqinfo)
}
