% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{get.p}
\alias{get.p}
\title{helper function for extracting pvalues from cor.assoc results}
\usage{
get.p(res)
}
\arguments{
\item{res}{Results of cor.assoc}
}
\value{
the matrix of the pvalues of the associations
}
\description{
helper function for extracting pvalues from cor.assoc results
}
