% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.s}
\alias{get.s}
\title{get.s}
\usage{
get.s(res)
}
\arguments{
\item{res}{Results of cor.assoc}
}
\value{
the matrix of associations
}
\description{
helper function for extracting statistics from cor.assoc results
}
\keyword{internal}
