% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{getMCE}
\alias{getMCE}
\title{helper function to calculate the MCE of the calibrations}
\usage{
getMCE(actual, predicted, n_bins = 10)
}
\arguments{
\item{actual}{real values of the variables}

\item{predicted}{predicted values by one of the surrogates}

\item{n_bins}{the number of bins}
}
\value{
MCE value
}
\description{
helper function to calculate the MCE of the calibrations
}
