% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getvol}
\alias{getvol}
\title{getvol}
\usage{
getvol()
}
\arguments{
\item{data}{the dataset to show}
}
\value{
table
}
\description{
helper function to retrieve the volumes for the download from the Rshiny application
}
\keyword{internal}
