% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{hist_plots_mortality}
\alias{hist_plots_mortality}
\title{Function to plot the histogram of the mortality score separated for different age ranges}
\usage{
hist_plots_mortality(mort_score, phenotypes)
}
\arguments{
\item{mort_score}{data.frame of the mortality score}

\item{phenotypes}{data.frame with age included in it}
}
\value{
plotly image with the histogram of the mortlity score separated in 3 age ranges
}
\description{
Function to plot the histogram of the mortality score separated for different age ranges
}
