% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.sym}
\alias{is.sym}
\title{is.sym}
\usage{
is.sym(res)
}
\arguments{
\item{res}{Results of cor.assoc}
}
\value{
TRUE/FALSE
}
\description{
Defining a helper function to check whether a supplied matrix is symmetric:
helper function to check whether a supplied matrix is symmetric
}
\keyword{internal}
