% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loading_spin}
\alias{loading_spin}
\title{loading_spin}
\usage{
loading_spin(plot)
}
\value{
loading spinner
}
\description{
helper function to create a loading spinner for the calibration
}
\keyword{internal}
