% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{metabo_names_translator}
\alias{metabo_names_translator}
\title{metabolomics feature nomenclatures}
\format{
An object of class \code{data.frame} with 228 rows and 9 columns.
}
\usage{
data("metabo_names_translator")
}
\description{
Translator of the names of the metabolomics-features to the ones used in BBMRI-nl
}
\examples{
data("metabo_names_translator")

}
\references{
This is a list originally created for the package ggforestplot and modified ad-hoc for our package
(https://nightingalehealth.github.io/ggforestplot/articles/index.html).
}
\keyword{datasets}
