% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{metabolites_subsets}
\alias{metabolites_subsets}
\title{metabolomics feature subsets}
\format{
An object of class \code{list} of length 8.
}
\usage{
data("metabolites_subsets")
}
\description{
List containing all the subset of the metabolomics-based features used for our models
}
\examples{
data("metabolites_subsets")

}
\references{
The selection of metabolic features available is the one selected by the papers:
Deelen,J. et al. (2019) A metabolic profile of all-cause mortality risk identified in an observational study of 44,168 individuals. Nature Communications, 10, 1-8, \url{doi:10.1038/s41467-019-11311-9}
Ahola-Olli,A.V. et al. (2019) Circulating metabolites and the risk of type 2 diabetes: a prospective study of 11,896 young adults from four Finnish cohorts. Diabetologia, 62, 2298-2309, \url{doi:10.1007/s00125-019-05001-w}
Wurtz,P. et al. (2015) Metabolite profiling and cardiovascular event risk: a prospective study of 3 population-based cohorts. Circulation, 131, 774-785, \url{doi:10.1161/CIRCULATIONAHA.114.013116}
Bizzarri,D. et al. (2022) 1H-NMR metabolomics-based surrogates to impute common clinical risk factors and endpoints. EBioMedicine, 75, 103764, \url{doi:10.1016/j.ebiom.2021.103764}
van den Akker Erik B. et al. (2020) Metabolic Age Based on the BBMRI-NL 1H-NMR Metabolomics Repository as Biomarker of Age-related Disease. Circulation: Genomic and Precision Medicine, 13, 541-547, \url{doi:10.1161/CIRCGEN.119.002610}
}
\keyword{datasets}
