% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{MetaboWAS}
\alias{MetaboWAS}
\title{Function that produces a MetaboWAS}
\usage{
MetaboWAS(met, pheno, test_variable, covariates, img = T)
}
\arguments{
\item{met}{metabolites data.frame}

\item{pheno}{phenotypes data.frame}

\item{test_variable}{the variable to be investigated}

\item{covariates}{the covariates that you want to add}

\item{img}{TRUE to plot a Manhattan plot}
}
\value{
res= the results of the MetaboWAS, manhplot= the Manhattan plot, N_hits= the number of significant hits
}
\description{
Function that produces a MetaboWAS
}
