% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{model_coeff_heat}
\alias{model_coeff_heat}
\title{Function to plot the scaled coefficients of the metabolic scores}
\usage{
model_coeff_heat(
  mort_betas,
  metaboAge_betas,
  surrogates_betas,
  Ahola_Olli_betas,
  CVD_score_betas,
  COVID_score_betas
)
}
\arguments{
\item{mort_betas}{dataframe withthe coefficients of the mortality score}

\item{metaboAge_betas}{dataframe with the coefficients of the metaboAge}

\item{surrogates_betas}{dataframe with the coefficients of the surrogates}

\item{Ahola_Olli_betas}{dataframe with the coefficients of the T2D score}

\item{CVD_score_betas}{dataframe with the coefficients of the CVD score}

\item{COVID_score_betas}{ataframe with the coefficients of the COVID_score}
}
\value{
heatmapply with the scaled coefficients of the metabolic scores
}
\description{
Function to plot the scaled coefficients of the metabolic scores
}
