% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MOLEPI_LCBC_header}
\alias{MOLEPI_LCBC_header}
\title{MOLEPI_LCBC_header}
\usage{
MOLEPI_LCBC_header()
}
\value{
header for Rshiny app
}
\description{
helper function to create a header with the links to MOLEPI, LCBC, LUMC and BBMRI-nl
}
\keyword{internal}
