% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{NA_message}
\alias{NA_message}
\title{helper function to create a plot indicating a problem with a plotly image}
\usage{
NA_message(main = "Metabolites are missing, please check your upload!")
}
\arguments{
\item{main}{Message to plot}
}
\value{
plot
}
\description{
helper function to create a plot indicating a problem with a plotly image
}
