\name{PARAM_metaboAge}
\alias{PARAM_metaboAge}
\docType{data}
\title{
Parameters to calculate MetaboAge
}
\description{
This list contains all the informations to calculate the MetaboAge by van den Akker et al.
}
\usage{data("PARAM_metaboAge")}
\format{
  The format is:
List of 8
 $ MET      : chr [1:56] "acace" "ace" "ala" "alb" ...
 $ Nmax_miss: num 1
 $ Nmax_zero: num 1
 $ logMEAN  : Named num [1:56] -3.37 -3.173 -1.233 -2.437 0.457 ...
  ..- attr(*, "names")= chr [1:56] "acace" "ace" "ala" "alb" ...
 $ logSD    : Named num [1:56] 0.747 0.391 0.259 0.084 0.126 ...
  ..- attr(*, "names")= chr [1:56] "acace" "ace" "ala" "alb" ...
 $ MEAN     : Named num [1:56] 0.0432 0.0445 0.3006 0.0877 1.5945 ...
  ..- attr(*, "names")= chr [1:56] "acace" "ace" "ala" "alb" ...
 $ SD       : Named num [1:56] 0.03465 0.01919 0.0767 0.00686 0.20025 ...
  ..- attr(*, "names")= chr [1:56] "acace" "ace" "ala" "alb" ...
 $ FIT_COEF : Named num [1:57] 58.619 1.056 0.689 -0.377 -2.843 ...
  ..- attr(*, "names")= chr [1:57] "(Intercept)" "acace" "ace" "ala" ...
}
\details{
 MET=Metabolites that are necessary for the logistic regression. Nmax_miss and Nmax_zero= the max number of missing and zeros raccommended for each sample. mean and sd= mean and standard deviation of each metabolites in BBMRI-NL. model_betas= a matrix with coefficients for the pre-trained models
}
\source{
https://www.ahajournals.org/doi/full/10.1161/CIRCGEN.119.002610
}
\examples{
data(PARAM_metaboAge)
## maybe str(PARAM_metaboAge) ; plot(PARAM_metaboAge) ...
}
\keyword{datasets}
