% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PARAM_surrogates}
\alias{PARAM_surrogates}
\title{PARAMETERS surrogates}
\format{
An object of class \code{list} of length 6.
}
\usage{
data("PARAM_surrogates")
}
\description{
The coefficients used to compute the metabolomics-based surrogate clinical variables by Bizzarri et al.
}
\details{
List containing all the information to pre-process and compute the surrogate clinical variables.
}
\examples{
data("PARAM_surrogates")

}
\references{
Bizzarri,D. et al. (2022) 1H-NMR metabolomics-based surrogates to impute common clinical risk factors and endpoints. EBioMedicine, 75, 103764, \url{doi:10.1016/j.ebiom.2021.103764}
}
\keyword{datasets}
