% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{pheno_barplots}
\alias{pheno_barplots}
\title{Function created to visualize the binarized variables in a barplot following the
rules described in Bizzarri et al.}
\usage{
pheno_barplots(bin_phenotypes)
}
\arguments{
\item{bin_phenotypes}{the binarized variables}
}
\value{
A plot with a barplot of the binarized variables
}
\description{
Function created to visualize the binarized variables in a barplot following the
rules described in Bizzarri et al.
}
