\name{phenotypes_names}
\alias{phenotypes_names}
\docType{data}
\title{
Nomenclatures for the phenotypes used in the app MiMIR
}
\description{
Phenotypes, Binary phenotypes,...
}
\usage{data("phenotypes_names")}
\format{
  The format is:
List of 5
 $ : chr [1:19] "sex" "diabetes" "lipidmed" "blood_pressure_lowering_med" ...
 $ : chr [1:19] "s_sex" "s_diabetes" "s_lipidmed" "s_blood_pressure_lowering_med" ...
 $ : chr [1:19] "sex" "diabetes" "lipidmed" "blood_pressure_lowering_med" ...
 $ : chr [1:19] "s_sex" "s_diabetes" "s_lipidmed" "s_blood_pressure_lowering_med" ...
 $ : chr [1:24] "sex" "diabetes" "lipidmed" "blood_pressure_lowering_med" ...
}
\examples{
data(phenotypes_names)
## maybe str(phenotypes_names) ; plot(phenotypes_names) ...
}
\keyword{datasets}
