% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{plattCalib_evaluation}
\alias{plattCalib_evaluation}
\title{Function that plots the Platt Calibrations using plotly}
\usage{
plattCalib_evaluation(
  r,
  p,
  name,
  nbins = 10,
  annot_x = c(1, 1),
  annot_y = c(0.1, 0.3)
)
}
\arguments{
\item{r}{binary real data}

\item{p}{predicted probabilities}

\item{name}{character string indicating the name of the calibrated variable}

\item{nbins}{number of bins to create the plots}

\item{annot_x}{integer indicating the x axis points in which the ECE and MCE values will be plotted}

\item{annot_y}{integer indicating the y axis points in which the ECE and MCE values will be plotted}
}
\value{
list with Reliability diagram and histogram with calibrations and original predictions
}
\description{
Function that plots the Platt Calibrations using plotly
}
