% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{plattCalib_plot}
\alias{plattCalib_plot}
\title{Function that plots the Platt Calibrations using plotly}
\usage{
plattCalib_plot(
  calibration,
  name,
  nbins = 10,
  annot_x = c(1.27, 1.27),
  annot_y = c(0.6, 0.4)
)
}
\arguments{
\item{calibration}{result of the plattCalibration function}

\item{name}{string with the name of the variable to be plotted}

\item{nbins}{number of bins to create the plots}

\item{annot_x}{integer indicating the x axis points in which the ECE and MCE values will be plotted}

\item{annot_y}{integer indicating the y axis points in which the ECE and MCE values will be plotted}
}
\value{
list with Reliability diagram and histogram with calibrations and original predictions
}
\description{
Function that plots the Platt Calibrations using plotly
}
