% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_cor_assoc}
\alias{plot_cor_assoc}
\title{Function for plotting dat1 x dat2 associations on basis of (partial) correlations:}
\usage{
plot_cor_assoc(
  res,
  main = NULL,
  zlim = NULL,
  reorder.x = FALSE,
  reorder.y = reorder.x,
  resort_on_p = FALSE,
  abs = FALSE,
  cor.abs = FALSE,
  text = TRUE,
  value = FALSE,
  cex.num = 8,
  fontsize = 8
)
}
\arguments{
\item{res}{associatinos obtained with cor.assoc}

\item{main}{title of the plot}

\item{zlim}{max association to plot}

\item{reorder.x}{TRUE/FALSE. If TRUE it will reorder the x axis based on clustering}

\item{reorder.y}{TRUE/FALSE. If TRUE it will reorder the y axis based on clustering}

\item{resort_on_p}{TRUE/FALSE. If TRUE it will reorder the x and y axis based on the pvalues of the associations}

\item{abs}{TRUE/FALSE. If TRUE it will reorder based the absolute values}

\item{cor.abs}{TRUE/FALSE. If TRUE it will plot the absolute values}

\item{text}{TRUE/FALSE. If TRUE it will include text in the plot}

\item{value}{TRUE/FALSE. If TRUE it will plot the values of the associations}

\item{cex.num}{integer indicating the size of the text}

\item{fontsize}{integer indicating the size of the title and axis}
}
\value{
heatmap with the results of cor.assoc
}
\description{
Function for plotting dat1 x dat2 associations on basis of (partial) correlations:
}
\keyword{internal}
