% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{plot_corply}
\alias{plot_corply}
\title{Function creating a plotly image with the dat1 x dat2 associations on basis of (partial) correlations:}
\usage{
plot_corply(
  res,
  main = NULL,
  zlim = NULL,
  reorder.x = FALSE,
  reorder.y = reorder.x,
  resort_on_p = FALSE,
  abs = FALSE,
  cor.abs = FALSE,
  reorder_dend = F
)
}
\arguments{
\item{res}{associatins obtained with cor.assoc}

\item{main}{title of the plot}

\item{zlim}{max association to plot}

\item{reorder.x}{TRUE/FALSE. If TRUE it will reorder the x axis based on clustering}

\item{reorder.y}{TRUE/FALSE. If TRUE it will reorder the y axis based on clustering}

\item{resort_on_p}{TRUE/FALSE. If TRUE it will reorder the x and y axis based on the pvalues of the associations}

\item{abs}{TRUE/FALSE. If TRUE it will reorder based the absolute values}

\item{cor.abs}{TRUE/FALSE. If TRUE it will plot the absolute values}

\item{reorder_dend}{TRUE/FALSE. If TRUE it will use the dendrogram calculated by plotly}
}
\value{
heatmap with the results of cor.assoc
}
\description{
Function creating a plotly image with the dat1 x dat2 associations on basis of (partial) correlations:
}
