% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{plot_na_heatmap}
\alias{plot_na_heatmap}
\title{Function for plotting a heatmap indcating missing & zero values}
\usage{
plot_na_heatmap(dat)
}
\arguments{
\item{dat}{The matrix or data.frame}
}
\value{
Plot with an heatmap and two histogram on the sides indicating zeros and missing values
}
\description{
Function for plotting a heatmap indcating missing & zero values
}
