% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{predictions_surrogates}
\alias{predictions_surrogates}
\title{Function that calculates one of the surrogate and creates figures with their accuracy}
\usage{
predictions_surrogates(FIT, data, title_img = FALSE, plot = TRUE)
}
\arguments{
\item{FIT}{The betas of the logistic regressions composing the surrogates by Bizzarri et al.}

\item{data}{The NH-metabolomics matrix}

\item{title_img}{string with title}

\item{plot}{TRUE/FALSE. If TRUE it calculates the roc curve}
}
\value{
If plot==T The surrogate predictions and the roc curve. If plot==F only the surrogate predictions
}
\description{
Function that calculates one of the surrogate and creates figures with their accuracy
}
