% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{prep.mort_data}
\alias{prep.mort_data}
\title{Function to prepare the metabolites data for Deelen et al. mortality score.}
\usage{
prep.mort_data(
  dat,
  featID = c("pufa_fa", "gp", "glc", "s_hdl_l", "xxl_vldl_l", "alb", "phe", "acace",
    "ile", "vldl_d", "leu", "val", "his", "lac"),
  quiet = FALSE
)
}
\arguments{
\item{dat}{The NH-metabolomics matrix}

\item{featID}{String vector with the names of metabolic features used}

\item{quiet}{TRUE/FALSE if TRUE if will suppress all messages from the function}
}
\value{
The NH-metabolomics matrix after checking for zeros and zscale the log metabolites concentrations
}
\description{
Function to prepare the metabolites data for Deelen et al. mortality score.
}
