% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{prep_met_for_scores}
\alias{prep_met_for_scores}
\title{Function to prepare the metabolites data for the mortality, T2D and CVD scores.}
\usage{
prep_met_for_scores(dat, featID, plusone = F, quiet = FALSE)
}
\arguments{
\item{dat}{The NH-metabolomics matrix}

\item{featID}{String vector with the names of metabolic features used}

\item{plusone}{TRUE or FALSe if 1.0 should be  added to all metabolites to calculate the score}

\item{quiet}{TRUE/FALSE if TRUE if will suppress all messages from the function}
}
\value{
The NH-metabolomics matrix after checking for zeros and zscale the log metabolites concentrations
}
\description{
Function to prepare the metabolites data for the mortality, T2D and CVD scores.
}
