% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileFixedFeatures.R
\name{profileFixedFeatures}
\alias{profileFixedFeatures}
\title{Calculate metadata profiles across fixed-width features}
\usage{
profileFixedFeatures(fixedRegions = NULL, sampleObject = NULL,
  TxDb = NULL, bins = c(150), weightCol = NULL,
  ignoreStrand = FALSE, dropEmpty = TRUE, normType = c("density",
  "max", "none"), collapse = TRUE, verbose = TRUE)
}
\arguments{
\item{fixedRegions}{A named \code{\link[GenomicRanges:GRangesList-class]{GenomicRanges}} or list of \code{\link[GenomicRanges:GRangesList-class]{GenomicRanges}} objects. The \code{\link[GenomicRanges:GRangesList-class]{GenomicRanges}} width must be the same within the same group, but it can vary between groups.}

\item{sampleObject}{A \code{\link[GenomicRanges:GRanges-class]{GenomicRanges}} or \code{\link[GenomicAlignments:GAlignments-class]{GenomicAlignments}} object.}

\item{TxDb}{A \code{\link[GenomicFeatures:TxDb-class]{GenomicFeatures}} object. Required if \code{fixedRegions} or \code{sampleObject} have missing \code{\link[GenomeInfoDb:Seqinfo-class]{GenomeInfoDb}} information. It must contain \code{\link[GenomeInfoDb:Seqinfo-class]{GenomeInfoDb}} information.}

\item{bins}{An ordered integer vector (must be greater equal that the length of \code{fixedRegions}). The vector order determines the number of bins for each region. If more bins than regions are provided, the additional will be ignored.}

\item{weightCol}{A single character string. This must be the name of an integer column in the \code{sampleObject} object.}

\item{ignoreStrand}{When set to 'TRUE', the strand information in \code{sampleObject} is ignored. This does not affect the features in \code{fixedRegions}.}

\item{dropEmpty}{When set to 'TRUE', the transcripts with no signal in any of their sub-regions will be discarded. When set to 'FALSE', all values of these regions will be set to 0.}

\item{normType}{A character string indicating which region normalising method to use. One of 'density' (default), 'max', 'none': can be abbreviated.
Depending on the chosen method the values of each region are normalised using the sum ('density'), the maximum ('max') of the values across the region, or not normalised at all ('none').}

\item{collapse}{When set to 'TRUE', the profiles are collapsed into a single profile.}

\item{verbose}{When set to 'TRUE', the function prints diagnostic messages.}
}
\value{
A \code{\link[data.table:data.table-class]{data.table}} of the normalised binned coverage across the fixed features. Column names are determined by \code{collapse}.
}
\description{
Calculate metadata profiles across fixed-width features
}
\examples{
library("TxDb.Dmelanogaster.UCSC.dm3.ensGene")
TxDb = TxDb.Dmelanogaster.UCSC.dm3.ensGene

query = genes(TxDb, filter=list(tx_chrom = "chr4"))
query = GRangesList(Promoter = promoters(query, upstream=500, downstream=1000))

library("pasillaBamSubset")
library("GenomicAlignments")

fl1 <- untreated1_chr4()
subject = readGAlignments(fl1)

profile = profileFixedFeatures(fixedRegions=query, sampleObject=subject, TxDb=TxDb)

library("ggplot2")

ggplot(profile, aes(x=bin, y=Mean)) + 
  geom_line() +
  geom_vline(xintercept=50, linetype="dashed", colour="grey30") +
  scale_x_continuous("Relative position",
       breaks=c(1, 50.5, 150), label=c("-500", "TSS", "1000")) +
  scale_y_continuous("Average normalised signal") +
  coord_cartesian(xlim=c(0, 150)) +
  theme_bw()
}
