% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileStrandedRegions.R
\name{profileStrandedRegions}
\alias{profileStrandedRegions}
\title{Calculate metadata profiles across GRanges or GRangesList}
\usage{
profileStrandedRegions(query, subject, nbin = 100)
}
\arguments{
\item{query}{A single stranded \code{\link[GenomicRanges:GenomicRanges-class]{GenomicRanges}} or \code{\link[GenomicRanges:GRangesList-class]{GenomicRanges}} objects, possibly named.}

\item{subject}{A \code{\link[IRanges:SimpleRleList-class]{IRanges}}, usually generated by the \code{\link[IRanges:coverage-methods]{coverage}} function.}

\item{nbin}{A single integer. It determines the number of bins in which each region will be divided.}
}
\value{
A \code{\link[data.table:data.table-class]{data.table}} of the normalised binned coverage across the genomic regions.
}
\description{
This function calculates coverage profiles across genomic regions.
}
\examples{
library("TxDb.Dmelanogaster.UCSC.dm3.ensGene")
TxDb = TxDb.Dmelanogaster.UCSC.dm3.ensGene

query = keepSeqlevels(exonsBy(TxDb, by="tx", use.names=TRUE), "chr4", pruning.mode="coarse")
query = query[strand(query)=="+"]
query = query[sum(width(query))>0]

library("pasillaBamSubset")
library("GenomicAlignments")

fl1 <- untreated1_chr4()
subject = coverage(keepSeqlevels(readGAlignments(fl1), "chr4"))

profile = profileStrandedRegions(query, subject)
profile = profile[, list(value = mean(value)), by="bin"]

library("ggplot2")

ggplot(profile, aes(x=bin, y=value)) + 
  geom_line() +
  scale_x_continuous("Position") +
  scale_y_continuous("Average normalised coverage") +
  theme_bw()
}
