% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{QCprep}
\alias{QCprep}
\title{Function to prepare the NH metabolomics dataset to compute MetaboAge by van den Akker et al.}
\usage{
QCprep(mat, PARAM_metaboAge, quiet = TRUE, Nmax_zero = 1, Nmax_miss = 1)
}
\arguments{
\item{mat}{The NH-metabolomics matrix; it may contain a mixture of flags and metabolites in the columns.}

\item{PARAM_metaboAge}{is a list holding the parameters to compute the metaboAge}

\item{quiet}{TRUE/FALSE if TRUE if will suppress all messages from the function}

\item{Nmax_zero}{max number of zeros allowed per sample}

\item{Nmax_miss}{max number of missing values allowed per sample}
}
\value{
The NH-metabolomics matrix after checking for zeros, missing values, samples>5SD from the BBMRI-mean, imputing the missing values and zscale metabolites concentrations
}
\description{
Function to prepare the NH metabolomics dataset to compute MetaboAge by van den Akker et al.
}
