% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{QCprep_surrogates}
\alias{QCprep_surrogates}
\title{Function to prepare the metabolomics dataset to compute the surrogates by Bizzarri et al.}
\usage{
QCprep_surrogates(
  mat,
  PARAM_surrogates,
  Nmax_miss = 1,
  Nmax_zero = 1,
  quiet = FALSE
)
}
\arguments{
\item{mat}{The NH-metabolomics matrix; it may contain a mixture of flags and metabolites in the columns.}

\item{PARAM_surrogates}{is a list holding the parameters to compute the surrogates}

\item{Nmax_miss}{max number of missing values allowed per sample}

\item{Nmax_zero}{max number of zeros allowed per sample}

\item{quiet}{TRUE/FALSE if TRUE if will suppress all messages from the function}
}
\value{
The NH-metabolomics matrix after checking for zeros, missing values, samples>5SD from the BBMRI-mean,
imputing the missing values and zscale metabolites concentrations
}
\description{
Function to prepare the metabolomics dataset to compute the surrogates by Bizzarri et al.
}
