% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rendertable}
\alias{rendertable}
\title{rendertable}
\usage{
rendertable(data)
}
\arguments{
\item{data}{the dataset to show}
}
\value{
table
}
\description{
helper function to create a table for an Rshiny app
}
\keyword{internal}
