% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{report.dim}
\alias{report.dim}
\title{Function to report on the console the dimension of the NH metabolomics matrix}
\usage{
report.dim(x, header, trailing = "0")
}
\arguments{
\item{x}{The NH metabolomics matrix}

\item{header}{string describing the subsampling of the NH-metabolomics matrix}

\item{trailing}{number of digits to include}
}
\value{
The report of the NH-metabolomics matrix dimension
}
\description{
Function to report on the console the dimension of the NH metabolomics matrix
}
