% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{resort.on.p}
\alias{resort.on.p}
\title{helper function for clustering cor.assoc results reordering based on the pvalues}
\usage{
resort.on.p(res, abs = FALSE)
}
\arguments{
\item{res}{Results of cor.assoc}

\item{abs}{TRUE/FALSE. If TRUE it will cluster the absolute values}
}
\value{
Returns the clustered order of the associations based on the pvalues
}
\description{
helper function for clustering cor.assoc results reordering based on the pvalues
}
