% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resort.on.s}
\alias{resort.on.s}
\title{resort.on.s}
\usage{
resort.on.s(res, abs = FALSE)
}
\arguments{
\item{res}{Results of cor.assoc}

\item{abs}{TRUE/FALSE. If TRUE it will cluster the absolute values}
}
\value{
Returns the clustered order of the associations
}
\description{
helper function for clustering cor.assoc results reordering based on the associations
}
\keyword{internal}
