% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{scatterplot_predictions}
\alias{scatterplot_predictions}
\title{scatterplot_predictions}
\usage{
scatterplot_predictions(x, p, title, xname, yname)
}
\arguments{
\item{x}{numeric vector}

\item{p}{second numeric vector}

\item{title}{string vector with the title}

\item{xname}{string vector with the name of the variable on the x axis}

\item{yname}{string vector with the name of the variable on the y axis}
}
\value{
plotly image with the scatterplot
}
\description{
Function to visualize a scatter-plot comparing two variables
}
\examples{
\dontrun{
library(plotly)
#load the metabolites dataset
metabolic_measures <- read.csv("Nightingale_file_path",header = TRUE, row.names = 1)
#load the phenotypic dataset
phenotypes <- read.csv("phenotypes_file_path",header = TRUE, row.names = 1)

#Pre-process the metabolic features
prepped_met<-QCprep(as.matrix(metabolic_measures, PARAM_metaboAge)
#Apply the metaboAge
metaboAge<-apply.fit(prepped_met, FIT=PARAM_metaboAge$FIT_COEF)

scatterplot_predictions(phenotypes$age, metaboAge, title="Chronological Age vs MetaboAge")
}

}
