% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{startApp}
\alias{startApp}
\title{Start metaboRiSc application.}
\usage{
startApp(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{TRUE/FALSE}
}
\value{
Opens application. If \code{launch.browser}=TRUE in default web browser
}
\description{
Start metaboRiSc application.
}
