% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{subset_samples_sd}
\alias{subset_samples_sd}
\title{Function created that subsets the NH-metabolomics matrix samples to the ones for which the metabolites
included in MetaboAge for which the log of the metabolic concentrations are not more than 5SD away from their mean}
\usage{
subset_samples_sd(x, MEAN, SD, quiet = FALSE)
}
\arguments{
\item{x}{The NH-metabolomics matrix}

\item{MEAN}{numeric vector indicating the mean of the metabolites in x}

\item{SD}{numeric vector indicating the standard deviations of the metabolites in x}

\item{quiet}{TRUE/FALSE. if TRUE if will suppress all messages from the function}
}
\value{
The NH-metabolomics matrix with the selected samples
}
\description{
Function created that subsets the NH-metabolomics matrix samples to the ones for which the metabolites
included in MetaboAge for which the log of the metabolic concentrations are not more than 5SD away from their mean
}
