% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{subset_samples_zero}
\alias{subset_samples_zero}
\title{Function that subsets the NH-metabolomics matrix samples to the ones with less than Nmax zeros}
\usage{
subset_samples_zero(x, Nmax = 1, quiet = FALSE)
}
\arguments{
\item{x}{The NH-metabolomics matrix}

\item{Nmax}{integer indicating  the max number of zeros allowed per sample}

\item{quiet}{TRUE/FALSE. if TRUE if will suppress all messages from the function}
}
\value{
The NH-metabolomics matrix with the selected samples
}
\description{
Function that subsets the NH-metabolomics matrix samples to the ones with less than Nmax zeros
}
