% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{synthetic_metabolic_dataset}
\alias{synthetic_metabolic_dataset}
\title{synthetic metabolomics dataset}
\format{
An object of class \code{data.frame} with 500 rows and 229 columns.
}
\usage{
data("synthetic_metabolic_dataset")
}
\description{
Data.frame containing a synthetic dataset of the Nightingale Metabolomics dataset created with the package synthpop from the LLS_PAROFF dataset.
}
\examples{
data("synthetic_metabolic_dataset")

}
\references{
M. Schoenmaker et al., 'Evidence of genetic enrichment for exceptional survival using a family approach: the Leiden Longevity Study', Eur. J. Hum. Genet., vol. 14, no. 1, Art. no. 1, Jan. 2006, \url{doi:10.1038/sj.ejhg.5201508}
B. Nowok, G. M. Raab, and C. Dibben, 'synthpop: Bespoke Creation of Synthetic Data in R', J. Stat. Softw., vol. 74, no. 1, Art. no. 1, Oct. 2016, \url{doi:10.18637/jss.v074.i11}
}
\keyword{datasets}
