\name{synthetic_phenotypic_dataset}
\alias{synthetic_phenotypic_dataset}
\docType{data}
\title{
Synthetic example phenotypic dataset made for the package MetaboRiSc
}
\description{
Synthetic example dataset made with synthpop
}
\usage{data("synthetic_phenotypic_dataset")}
\format{
  A data frame with 5000 observations on the following 20 variables.
  \describe{
    \item{\code{sex}}{a numeric vector}
    \item{\code{diabetes}}{a numeric vector}
    \item{\code{lipidmed}}{a numeric vector}
    \item{\code{blood_pressure_lowering_med}}{a numeric vector}
    \item{\code{current_smoking}}{a numeric vector}
    \item{\code{metabolic_syndrome}}{a logical vector}
    \item{\code{alcohol_consumption}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{BMI}}{a numeric vector}
    \item{\code{ln_hscrp}}{a numeric vector}
    \item{\code{waist_circumference}}{a logical vector}
    \item{\code{weight}}{a numeric vector}
    \item{\code{height}}{a numeric vector}
    \item{\code{triglycerides}}{a numeric vector}
    \item{\code{ldl_chol}}{a numeric vector}
    \item{\code{hdlchol}}{a numeric vector}
    \item{\code{totchol}}{a numeric vector}
    \item{\code{eGFR}}{a numeric vector}
    \item{\code{wbc}}{a numeric vector}
    \item{\code{hgb}}{a numeric vector}
  }
}
\examples{
data(synthetic_phenotypic_dataset)
## maybe str(synthetic_phenotypic_dataset) ; plot(synthetic_phenotypic_dataset) ...
}
\keyword{datasets}
