% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{acc_LOBOV}
\alias{acc_LOBOV}
\title{acc_LOBOV}
\format{
An object of class \code{list} of length 20.
}
\usage{
data("acc_LOBOV")
}
\description{
Accuracy of the Leave One Biobank Out Validation of the surrogate metabolic-modesl performed in BBMRI-nl
}
\details{
Dataframe containing the accuracy obtained during the Leave One Biobank Out Validation of the surrogate metabolic-modesl  in BBMRI-nl.
}
\examples{
data("acc_LOBOV")

}
\references{
The method is described in:
Bizzarri,D. et al. (2022) 1H-NMR metabolomics-based surrogates to impute common clinical risk factors and endpoints. EBioMedicine, 75, 103764, \url{doi:10.1016/j.ebiom.2021.103764}
}
\keyword{datasets}
