% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{apply.fit_surro}
\alias{apply.fit_surro}
\title{Function that apply on of the surrogates models to the NH-metabolomics concentrations}
\usage{
apply.fit_surro(mat, FIT, post = TRUE)
}
\arguments{
\item{mat}{The NH-metabolomics matrix}

\item{FIT}{The betas of the logistic regressions composing the surrogates by Bizzarri et al.}

\item{post}{True if you want to obtain posterior probabilities}
}
\value{
The surrogates by Bizzarri et al.
}
\description{
Function that apply on of the surrogates models to the NH-metabolomics concentrations
}
