% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{apply.scale}
\alias{apply.scale}
\title{Function created to scale the NH-metabolomics matrix samples}
\usage{
apply.scale(dat, MEAN, SD, quiet = TRUE)
}
\arguments{
\item{dat}{The matrix}

\item{MEAN}{numeric vector indicating the mean of the metabolites present in dat}

\item{SD}{numeric vector indicating the standard deviations of the metabolites present in dat}

\item{quiet}{TRUE/FALSE. if TRUE if will suppress all messages from the function}
}
\value{
The matrix z-scaled using the means and sds given
}
\description{
Function created to scale the NH-metabolomics matrix samples
}
