% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BBMRI_hist_scaled}
\alias{BBMRI_hist_scaled}
\title{BBMRI_hist_scaled}
\format{
An object of class \code{list} of length 57.
}
\usage{
data("BBMRI_hist_scaled")
}
\description{
Z-scaled distributions of the Nightingale Health metabolic features in BBMRI-nl
}
\details{
List containing the histograms of the scaled metabolomics-features in BBMRI-nl.
}
\examples{
data("BBMRI_hist_scaled")

}
\keyword{datasets}
