\name{c21}
\alias{c21}
\docType{data}
\title{
Colors related to each score
}
\description{
A data.frame containing colors related to the scores for the plots.
}
\usage{data("c21")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\examples{
data(c21)
## maybe str(c21) ; plot(c21) ...
}
\keyword{datasets}
