% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{comp.CVD_score}
\alias{comp.CVD_score}
\title{Function to compute CVD score by Peter Wurtz on metabolite data.}
\usage{
comp.CVD_score(met, phen, betas = CVD_score_betas, quiet = FALSE)
}
\arguments{
\item{met}{The NH-metabolomics matrix of the samples}

\item{phen}{phenotypic information of the samples}

\item{betas}{The betas of the linear regression composing the score}

\item{quiet}{TRUE/FALSE if TRUE if will suppress all messages from the function}
}
\value{
The CVD score by Peter Wurtz
}
\description{
Function to compute CVD score by Peter Wurtz on metabolite data.
}
