% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{comp.T2D_Ahola_Olli}
\alias{comp.T2D_Ahola_Olli}
\title{Function to compute T2D score by Ahola Olli mortality score on metabolite data.}
\usage{
comp.T2D_Ahola_Olli(met, phen, betas = Ahola_Olli_betas, quiet = FALSE)
}
\arguments{
\item{betas}{The betas of the linear regression composing the mortality score}

\item{quiet}{TRUE/FALSE if TRUE if will suppress all messages from the function}

\item{dat}{The NH-metabolomics matrix}
}
\value{
The T2D by Ahola Olli
}
\description{
Function to compute T2D score by Ahola Olli mortality score on metabolite data.
}
