% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{cor_assoc}
\alias{cor_assoc}
\title{SubFunction for plotting data1 x data2 associations on basis of (partial) correlations:}
\usage{
cor_assoc(dat1, dat2, feat1, feat2, method = "pearson", quiet = FALSE)
}
\arguments{
\item{dat1}{matrix 1}

\item{dat2}{matrix 2}

\item{feat1}{string with the names of the selected variables in dat}

\item{feat2}{string with the names of the selected variables in dat2}

\item{method}{indicates which methods of the correlation to use (look at cor.test)}

\item{quiet}{TRUE/FALSE. if TRUE if will suppress all messages from the function}
}
\value{
correlations of the selected variables in the 2 martrices
}
\description{
SubFunction for plotting data1 x data2 associations on basis of (partial) correlations:
}
