% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{find_BBMRI_names}
\alias{find_BBMRI_names}
\title{Function to translate metabolites names alternatives to the BBMRI names}
\usage{
find_BBMRI_names(names)
}
\arguments{
\item{names}{The NH-metabolomics names}
}
\value{
The uploaded names and the BBMRI names of the metabolites available in BBMRI.
}
\description{
Function to translate metabolites names alternatives to the BBMRI names
}
