% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{get.s}
\alias{get.s}
\title{helper function for extracting statistics from cor.assoc results}
\usage{
get.s(res)
}
\arguments{
\item{res}{Results of cor.assoc}
}
\value{
the matrix of associations
}
\description{
helper function for extracting statistics from cor.assoc results
}
