% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getECE}
\alias{getECE}
\title{getECE}
\usage{
getECE(actual, predicted, n_bins = 10)
}
\arguments{
\item{actual}{observed binary phenotype}

\item{predicted}{predicted values}

\item{n_bins}{the number of bins}
}
\value{
ECE value
}
\description{
helper function to calculate the ECE of calibrations
}
\keyword{internal}
