% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{hist_plots}
\alias{hist_plots}
\title{Function to plot the histogram of the selected variables}
\usage{
hist_plots(
  dat,
  x_name,
  color = c21,
  scaled = FALSE,
  main = "Predictors Distributions"
)
}
\arguments{
\item{dat}{data.frame or matrix with the variables to plot}

\item{x_name}{string with the names of the selected variables in dat}

\item{color}{colors selected for all the variables}

\item{scaled}{TRUE/FALSE. If TRUE the variables will be z-scaled before plotting them}

\item{main}{title of the plot}
}
\value{
plotly image with the histograms of the selected variables
}
\description{
Function to plot the histogram of the selected variables
}
