% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{impute.miss}
\alias{impute.miss}
\title{Function that serves at imputing the missing values with zeros}
\usage{
impute.miss(x)
}
\arguments{
\item{x}{The matrix}
}
\value{
The matrix with missing values imputed to zero
}
\description{
Function that serves at imputing the missing values with zeros
}
