% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{is.sym}
\alias{is.sym}
\title{helper function to check whether a supplied matrix is symmetric}
\usage{
is.sym(res)
}
\arguments{
\item{res}{Results of cor.assoc}
}
\value{
TRUE/FALSE
}
\description{
helper function to check whether a supplied matrix is symmetric
}
