% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{kapmeier_scores}
\alias{kapmeier_scores}
\title{Function that creates a Kaplan Meier for an Event of the predictor divided by its mean}
\usage{
kapmeier_scores(predictors, pheno, score, Eventname = "Event")
}
\arguments{
\item{predictors}{The data.frame containing the predictors}

\item{pheno}{The data.frame containing the phenotypes}

\item{score}{a character string indicating which predictor to use}

\item{Eventname}{a character string with the name of the event to print on the plot}
}
\value{
plotly with a Kaplan Meier for an Event of the predictor divided by its mean
}
\description{
Function that creates a Kaplan Meier for an Event of the predictor divided by its mean
}
