% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{LOBOV_accuracies}
\alias{LOBOV_accuracies}
\title{Function created to visualize the accuracies in the current dataset compared to the
accuracies in the Leave One Biobank Out Validation in Bizzarri et al.}
\usage{
LOBOV_accuracies(surrogates, bin_phenotypes, bin_pheno_available, acc_LOBOV)
}
\arguments{
\item{surrogates}{matrix with the surrogate variables calculated in Bizzarri et all.}

\item{bin_phenotypes}{magtrix with the binarized variables}

\item{bin_pheno_available}{character vector with the phenotypes availables}

\item{acc_LOBOV}{accuracy of LOBOV calculated in Bizzarri et al.}
}
\value{
Boxplot with the accuracies of the LOBOV
}
\description{
Function created to visualize the accuracies in the current dataset compared to the
accuracies in the Leave One Biobank Out Validation in Bizzarri et al.
}
