\name{metabolites_subsets}
\alias{metabolites_subsets}
\docType{data}
\title{
The subsets of the Nightingale metabolomics platform used in the application MiMIR
}
\description{
Subsets of the metabolites for the app
}
\usage{data("metabolites_subsets")}
\format{
  The format is:
List of 8
 $ : chr [1:63] "ala" "gln" "his" "phe" ...
 $ : chr [1:62] "ala" "gln" "his" "phe" ...
 $ : chr [1:57] "ala" "gln" "his" "phe" ...
 $ : chr [1:56] "ala" "gln" "his" "phe" ...
 $ : chr [1:14] "pufa_fa" "gp" "glc" "s_hdl_l" ...
 $ : chr [1:25] "gp" "dha" "crea" "mufa" ...
 $ : chr [1:3] "phe" "l_vldl_ce_percentage" "l_hdl_fc"
 $ : chr [1:4] "phe" "mufa_fa" "faw6" "dha"
}
\examples{
data(metabolites_subsets)
## maybe str(metabolites_subsets) ; plot(metabolites_subsets) ...
}
\keyword{datasets}
