% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mort_betas}
\alias{mort_betas}
\title{Mortality score betas}
\format{
An object of class \code{data.frame} with 14 rows and 3 columns.
}
\usage{
data("mort_betas")
}
\description{
The coefficients used to compute the mortality score by Deelen et al.
}
\details{
Dataframe containing the abbreviation of the metabolites, the metabolites names and finally the Coefficients to compute the mortality score
}
\examples{
data("mort_betas")

}
\references{
Deelen,J. et al. (2019) A metabolic profile of all-cause mortality risk identified in an observational study of 44,168 individuals. Nature Communications, 10, 1-8, \url{doi:10.1038/s41467-019-11311-9}
}
\keyword{datasets}
