% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{multi_hist}
\alias{multi_hist}
\title{multi_hist}
\usage{
multi_hist(dat, color = MiMIR::c21, scaled = FALSE)
}
\arguments{
\item{dat}{data.frame or matrix with the variables to plot}

\item{color}{colors selected for all the variables}

\item{scaled}{logical to z-scale the variables}
}
\value{
plotly image with the histograms for all the variables in dat
}
\description{
#' Function to plot the histograms for all the variables in dat
}
\examples{
\dontrun{
library(plotly)
library(MiMIR)

#load the metabolites dataset
metabolic_measures <- read.csv("Nightingale_file_path",header = TRUE, row.names = 1)

multi_hist(metabolic_measures, scaled=T)
}

}
