% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{multi_hist}
\alias{multi_hist}
\title{Function to plot the histograms for all the variables in dat}
\usage{
multi_hist(dat, color = MiMIR::c21, scaled = FALSE)
}
\arguments{
\item{dat}{data.frame or matrix with the variables to plot}

\item{color}{colors selected for all the variables}

\item{scaled}{TRUE/FALSE. If TRUE the variables will be z-scaled}
}
\value{
plotly image with the histograms for all the variables in dat
}
\description{
Function to plot the histograms for all the variables in dat
}
