\name{PARAM}
\alias{PARAM}
\docType{data}
\title{ 
Parameters to calculate MetaboAge
}
\description{
This list contains all the informations to calculate the MetaboAge by van den Akker et all.
}
\usage{data("PARAM")}
\format{
  The format is:
List of 8
 $ MET      : chr [1:56] "acace" "ace" "ala" "alb" ...
 $ Nmax_miss: num 1
 $ Nmax_zero: num 1
 $ logMEAN  : Named num [1:56] -3.37 -3.173 -1.233 -2.437 0.457 ...
  ..- attr(*, "names")= chr [1:56] "acace" "ace" "ala" "alb" ...
 $ logSD    : Named num [1:56] 0.747 0.391 0.259 0.084 0.126 ...
  ..- attr(*, "names")= chr [1:56] "acace" "ace" "ala" "alb" ...
 $ MEAN     : Named num [1:56] 0.0432 0.0445 0.3006 0.0877 1.5945 ...
  ..- attr(*, "names")= chr [1:56] "acace" "ace" "ala" "alb" ...
 $ SD       : Named num [1:56] 0.03465 0.01919 0.0767 0.00686 0.20025 ...
  ..- attr(*, "names")= chr [1:56] "acace" "ace" "ala" "alb" ...
 $ FIT_COEF : Named num [1:57] 58.619 1.056 0.689 -0.377 -2.843 ...
  ..- attr(*, "names")= chr [1:57] "(Intercept)" "acace" "ace" "ala" ...
}
\details{
 MET=Metabolites that are necessary for the linear regression. Nmax_miss and Nmax_zero= the max number of missing and zeros raccommended for each sample. logMEAN = the  mean of each of the log transformed metabolites in BBMRI-NL. lodSD =the  standard deviation of each of the log transformed metabolites in BBMRI-NL. MEAN and SD= mean and standard deviation of each metabolites in BBMRI-NL. FIT_COEF= the coefficients for the pre-trained model
}
\source{
van den Akker Erik B. et al., ‘Metabolic Age Based on the BBMRI-NL 1H-NMR Metabolomics Repository as Biomarker of Age-related Disease’, Circ. Genomic Precis. Med., vol. 13, no. 5, pp. 541–547, Oct. 2020, doi: 10.1161/CIRCGEN.119.002610.
}
\examples{
data(PARAM)
## maybe str(PARAM) ; plot(PARAM) ...
}
\keyword{datasets}
