% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PARAM_metaboAge}
\alias{PARAM_metaboAge}
\title{PARAMETERS MetaboAge}
\format{
An object of class \code{list} of length 8.
}
\usage{
data("PARAM_metaboAge")
}
\description{
The coefficients used to compute the MetaboAge by van den Akker et al.
}
\details{
List containing all the information to pre-process and compute the MetaboAge.
}
\examples{
data("PARAM_metaboAge")

}
\references{
van den Akker Erik B. et al. (2020) Metabolic Age Based on the BBMRI-NL 1H-NMR Metabolomics Repository as Biomarker of Age-related Disease. Circulation: Genomic and Precision Medicine, 13, 541-547, \url{doi:10.1161/CIRCGEN.119.002610}
}
\keyword{datasets}
