% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{phenotypes_names}
\alias{phenotypes_names}
\title{phenotypic features names}
\format{
An object of class \code{list} of length 5.
}
\usage{
data("phenotypes_names")
}
\description{
List containing all the subsets of phenotypics variables used in the app
}
\examples{
data("phenotypes_names")

}
\keyword{datasets}
