% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{plattCalibration}
\alias{plattCalibration}
\title{Function that calculates the Platt Calibrations}
\usage{
plattCalibration(r.calib, p.calib, nbins = 10, plot = FALSE)
}
\arguments{
\item{r.calib}{binary calibration data response}

\item{p.calib}{numeric calibration predicted probabilities}

\item{nbins}{number of bins to create the plots}

\item{plot}{TRUE/FALSE. If TRUE it will plot the Reliability diagram and histogram of the calibrations}
}
\value{
list with samples, responses, calibrations, ECE, MCE and calibration plots if save==T
}
\description{
Function that calculates the Platt Calibrations
}
