% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{plot_na_heatmap}
\alias{plot_na_heatmap}
\title{plot_na_heatmap}
\usage{
plot_na_heatmap(dat)
}
\arguments{
\item{dat}{The matrix or data.frame}
}
\value{
Plot with a central heatmap and two histogram on the sides
}
\description{
Function plotting information about missing & zero values on the indicated matrix.
}
\details{
This heatmap indicates the available values in grey and missing or zeros in white.
On the sides two bar plots on the sides, one showing the missingn or zero values
per row and another to show the missing or zeroes per column.
}
\examples{
library(graphics)
library(MiMIR)

#load the metabolites dataset
metabolic_measures <- synthetic_metabolic_dataset
#Plot the missing values in the metabolomics matrix
plot_na_heatmap(metabolic_measures)

}
