% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotly_NA_message}
\alias{plotly_NA_message}
\title{plotly_NA_message}
\usage{
plotly_NA_message(main = "Phenotype not available!")
}
\arguments{
\item{main}{Message to plot}
}
\value{
plotly image
}
\description{
helper function to create a plotly indicating a problem with a plotly image
}
\keyword{internal}
