% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{prep_data_COVID_score}
\alias{prep_data_COVID_score}
\title{Function to prepare the metabolites data for the COVID score}
\usage{
prep_data_COVID_score(
  dat,
  featID = c("gp", "dha", "crea", "mufa", "apob_apoa1", "tyr", "ile", "sfa_fa", "glc",
    "lac", "faw6_faw3", "phe", "serum_c", "faw6_fa", "ala", "pufa", "glycine", "his",
    "pufa_fa", "val", "leu", "alb", "faw3", "ldl_c", "serum_tg"),
  quiet = FALSE
)
}
\arguments{
\item{dat}{The NH-metabolomics matrix}

\item{featID}{String vector with the names of metabolic features used}

\item{quiet}{TRUE/FALSE if TRUE if will suppress all messages from the function}
}
\value{
The NH-metabolomics matrix after checking for zeros and zscale the log metabolites concentrations
}
\description{
Function to prepare the metabolites data for the COVID score
}
