% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{roc_surro}
\alias{roc_surro}
\title{Function that creates a plotly image of the selected surrogates}
\usage{
roc_surro(surrogates, bin_phenotypes, x_name)
}
\arguments{
\item{surrogates}{The data.frame containing the surrogate values by Bizzarri et al.}

\item{bin_phenotypes}{data.frame with the binarized phenotypes}

\item{x_name}{string with the names of the selected binary phenotypes for the roc}
}
\value{
plotly image with the ROC curves for the selected variables
}
\description{
Function that creates a plotly image of the selected surrogates
}
