% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{roc_surro_subplots}
\alias{roc_surro_subplots}
\title{Function that creates a plotly image with the subplots with ROCs of the surrogates of the available variables}
\usage{
roc_surro_subplots(surrogates, bin_phenotypes)
}
\arguments{
\item{surrogates}{data.frame containing the surrogate values by Bizzarri et al.}

\item{bin_phenotypes}{data.frame with the binarized phenotypes output of binarize_all_pheno}
}
\value{
plotly image with all the ROCs for the available clinical variables
}
\description{
Function that creates a plotly image with the subplots with ROCs of the surrogates of the available variables
}
