% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{scatterplot_predictions}
\alias{scatterplot_predictions}
\title{Function that creates a scatterplot with the real values and predicted values}
\usage{
scatterplot_predictions(x, p, title, xname, yname)
}
\arguments{
\item{x}{numeric vector with the real values}

\item{p}{numeric vector with the predicted values}

\item{title}{string with the title}

\item{xname}{string with the name of the variable on the x axis}

\item{yname}{string with the name of the variable on the y axis}
}
\value{
plotly image with the scatterplot of a real vs predicted variable
}
\description{
Function that creates a scatterplot with the real values and predicted values
}
