% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{subset_metabolites_overlap}
\alias{subset_metabolites_overlap}
\title{Function that subsets the NH-metabolomics matrix features to the selection in metabos}
\usage{
subset_metabolites_overlap(x, metabos, quiet = FALSE)
}
\arguments{
\item{x}{The NH-metabolomics matrix}

\item{metabos}{String vector containing the metabolites concentration names to be selected}

\item{quiet}{TRUE/FALSE if TRUE if will suppress all messages from the function}
}
\value{
The NH-metabolomics matrix with the selected metabolites
}
\description{
Function that subsets the NH-metabolomics matrix features to the selection in metabos
}
