\name{synthetic_phenotypic_dataset}
\alias{synthetic_phenotypic_dataset}
\docType{data}
\title{
Synthetic example phenotypic dataset made for the package MetaboRiSc
}
\description{
Synthetic example dataset made with synthpop
}
\usage{data("synthetic_phenotypic_dataset")}
\format{
  A data frame with 500 observations on the following 20 variables.
}
\examples{
data(synthetic_phenotypic_dataset)
## maybe str(synthetic_phenotypic_dataset) ; plot(synthetic_phenotypic_dataset) ...
}
\keyword{datasets}
