% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{ttest_scores}
\alias{ttest_scores}
\title{Function that creates a boxplot with a continuous variable split using the binary variable}
\usage{
ttest_scores(dat, pred, pheno)
}
\arguments{
\item{dat}{The data.frame containing the 2 variables}

\item{pred}{character indicating the y variable}

\item{pheno}{character indicating the binary variable}
}
\value{
plotly boxplot with the continuous variable split using the binary variable
}
\description{
Function that creates a boxplot with a continuous variable split using the binary variable
}
