% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_functions.R
\name{ttest_surrogates}
\alias{ttest_surrogates}
\title{Function that creates a plotly image of the selected surrogates}
\usage{
ttest_surrogates(surrogates, bin_phenotypes)
}
\arguments{
\item{surrogates}{The data.frame containing the surrogate values by Bizzarri et al.}

\item{bin_phenotypes}{data.frame with the binarized phenotypes}
}
\value{
plotly image with the ROC curves for the selected variables
}
\description{
Function that creates a plotly image of the selected surrogates
}
